### 2024-10-22 #7393 @editorialbot
[PRE REVIEW]: gaia: An R package to estimate crop yield responses to temperature and precipitation
**Submitting author:** @mengqi-z (<a href="http://orcid.org/0000-0001-5385-2758">Mengqi Zhao</a>)
**Repository:** https://github.com/JGCRI/gaia
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.0
