### 2024-10-24 #7393 @kthyng
@mengqi-z
* in the docs, a suggestion I have is to add what the user can expect to look for for success from the examples, like certain values, outcomes, and/or especially a plot they can make to confirm results.
* Please reduce to length of your paper to up to about 1000 words. Paper format information here: https://joss.readthedocs.io/en/latest/paper.html. You can check paper length (and a few other items) with the command `@editorialbot check repository` though note it needs to be the first item in a comment to run correctly.
