### 2024-10-28 45fdf03
Miscellaneous Updates
1. Update to use the correct historical observations to fit the model. Zenodo link in the vignette is updated to get the correct historical data. Updated corresponding testthat.
2. Update the user guide to include more examples and details of the outputs.
