+++ vignettes/vignette.R
@@ -1,43 +1,70 @@
+## ----setup, include=FALSE-----------------------------------------------------
+knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
+
 ## ----eval=F-------------------------------------------------------------------
 #  
 #  # load gaia
 #  library(gaia)
 #  
-#  # NOTE: please change `data_dir` to your desired location for downloaded data
+#  # Path to the output folder. Change it to your desired location
+#  output_dir <- file.path(getwd(), 'gaia_output')
+#  
+#  # Cropland-weighted historical and future climate data
 #  data_dir <- gaia::get_example_data(
-#    download_url = 'https://zenodo.org/records/13179630/files/gaea_example_climate.zip?download=1',
-#    data_dir = 'path/to/desired/location'
+#    download_url = 'https://zenodo.org/records/13976521/files/weighted_climate.zip?download=1',
+#    data_dir = output_dir
 #  )
 #  
-#  # Path to the climate NetCDF files
-#  # NOTE: Each variable can have more than one file
-#  # historical climate data
-#  pr_historical_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_rcp7_1950_2014.nc')
-#  tas_historical_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_rcp7_1950_2014.nc')
-#  
-#  # projected climate data
-#  pr_projection_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_rcp7_2015_2100.nc')
-#  tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_rcp7_2015_2100.nc')
+#  # Path to the folder that holds cropland-area-weighted precipitation and temperature TXT files
+#  # historical climate observation
+#  climate_hist_dir <- file.path(data_dir, 'climate_hist')
+#  # future projected climate
+#  climate_impact_dir <- file.path(data_dir, 'canesm5')
 #  
 
-## ----eval=FALSE---------------------------------------------------------------
+## ----eval=F-------------------------------------------------------------------
 #  
 #  # load gaia
 #  library(gaia)
+#  
+#  # Path to the output folder. Change it to your desired location
+#  output_dir <- file.path(getwd(), 'gaia_output')
+#  
+#  # Future Climate Data
+#  data_dir <- gaia::get_example_data(
+#    download_url = 'https://zenodo.org/records/13976521/files/gaia_example_climate.zip?download=1',
+#    data_dir = output_dir
+#  )
+#  
+#  # Path to the precipitation and temperature NetCDF files
+#  # NOTE: Each variable can have more than one file
+#  # projected climate data
+#  pr_projection_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_gcam-ref_2015_2100.nc')
+#  tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_gcam-ref_2015_2100.nc')
+#  
+
+## ----eval=F, echo=T-----------------------------------------------------------
+#  
+#  # load gaia
+#  library(gaia)
+#  
+#  # Path to the output folder. Change it to your desired location
+#  output_dir <- file.path(getwd(), 'gaia_output')
 #  
 #  # Run gaia
 #  # The full run with raw climate data can take up to an hour
 #  gaia::yield_impact(
-#    pr_hist_ncdf = pr_historical_file,
-#    tas_hist_ncdf = tas_historical_file,
-#    pr_proj_ncdf = pr_projection_file,
-#    tas_proj_ncdf = tas_projection_file,
+#    pr_hist_ncdf = NULL,                    # path to historical precipitation NetCDF file (must follow ISIMIP format); only if you wish to use your own historical precipitation observation
[…]
