+++ README.md
@@ -18,7 +18,7 @@
 <br />
 
 <p align="center">
-<a href="https://jgcri.github.io/gaia/" target="_blank"><img src="https://github.com/JGCRI/jgcricolors/blob/main/vignettes/button_user_guide.PNG?raw=true" 
+<a href="https://jgcri.github.io/gaia/" target="_blank"><img src="https://github.com/JGCRI/gaia/blob/main/vignettes/vignetteFigs/button_user_guide.png?raw=true" 
 alt="https://jgcri.github.io/gaia/articles/vignette.html" height="60"/></a>
 <img src="https://github.com/JGCRI/jgcricolors/blob/main/vignettes/button_divider.PNG?raw=true" height="40"/>
 </p>

[…]
