### 2024-10-28 bc46a9d
Update APG figure
