+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   vignette: vignette.html
-last_built: 2024-10-28T20:08Z
+last_built: 2024-10-28T20:10Z
