### 2024-10-28 0abab29
Update paper.bib to fix few DOIs  and modify workflow name
