+++ paper/paper.bib
@@ -34,7 +34,7 @@ volume = {24},
 number = {1},
 pages = {},
 keywords = {land use, irrigation, agricultural systems},
-doi = {https://doi.org/10.1029/2008GB003435},
+doi = {10.1029/2008GB003435},
 url = {https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2008GB003435},
 eprint = {https://agupubs.onlinelibrary.wiley.com/doi/pdf/10.1029/2008GB003435},
 year = {2010}
@@ -60,7 +60,7 @@ volume = {279},
 pages = {108448},
 year = {2022},
 issn = {0378-4290},
-doi = {https://doi.org/10.1016/j.fcr.2022.108448},
+doi = {10.1016/j.fcr.2022.108448},
 url = {https://www.sciencedirect.com/science/article/pii/S0378429022000193},
 author = {Ranju Chapagain and Tomas A. Remenyi and Rebecca M.B. Harris and Caroline L. Mohammed and Neil Huth and Daniel Wallach and Ehsan Eyshi Rezaei and Jonathan J. Ojeda},
 keywords = {Crop model, Input uncertainty, Structural uncertainty, Parameter uncertainty, Grain yield, Mechanistic modelling}
@@ -100,7 +100,7 @@ volume = {19},
 number = {5},
 pages = {607-620},
 keywords = {Agricultural management, crop calendars, crop modelling, global agricultural systems, harvesting dates, maize, phenology, planting dates, spring wheat, winter wheat},
-doi = {https://doi.org/10.1111/j.1466-8238.2010.00551.x},
+doi = {10.1111/j.1466-8238.2010.00551.x},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/j.1466-8238.2010.00551.x},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/j.1466-8238.2010.00551.x},
 year = {2010}
@@ -161,7 +161,7 @@ volume = {45},
 number = {1},
 pages = {3-20},
 keywords = {Computable general equilibrium, Partial equilibrium, Meta-analysis, Socioeconomic pathway, Climate change, Bioenergy, Land use, Model intercomparison, C63, C68, Q11, Q16, Q24, Q42, Q54},
-doi = {https://doi.org/10.1111/agec.12086},
+doi = {10.1111/agec.12086},
 url = {https://onlinelibrary.wiley.com/doi/abs/10.1111/agec.12086},
 eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1111/agec.12086},
 year = {2014}
