+++ .github/workflows/R-CDM-check.yaml
@@ -1,4 +1,4 @@
-name: R-CMD-check.yaml
+name: R-CMD-check
 
 on:
   push:
