### 2024-11-04 17252f9
Update vignette and default coeff
