### 2024-11-04 e4f2259
Update docs.yaml
maxOS-12 is depreciated
