### 2024-11-04 a7e87f4
Fix the ggplot2 function name
