### 2024-11-04 8082f22
test macos-13 in docs.yaml
there seems to be an issue with macos check with data.table.
