+++ .github/workflows/docs.yaml
@@ -8,7 +8,7 @@ name: docs
 
 jobs:
   pkgdown:
-    runs-on: macos-latest
+    runs-on: macos-13
     env:
       GITHUB_PAT: ${{ secrets.GAEA_PAT }}
     steps:
