+++ .github/workflows/docs.yaml
@@ -1,8 +1,8 @@
 on:
   push:
-    branches:
-      - main
-      - dev
+    branches: [main]
+  pull_request:
+    branches: [main]
 
 name: docs
 
