### 2024-11-05 #7393 @mengqi-z
@kthyng - Apologies for the delay, and thank you for your suggestions. I have made the following updates:

- [ ] Expanded the tutorial to include more detailed examples, along with some expected outputs from `gaia`.
- [ ] Revised the paper to reduce the word count from Summary Section to Acknowledgement Section to around 1025. I didn't count the words from title, authors, or institutions.
