+++ R/gcam_agprodchange.R
@@ -330,7 +330,8 @@ get_weighted_yield_impact <- function(data = NULL,
   # ----------------------------------------------------------------------------
 
 
-  years <- paste0('X', c(2015, seq(2020, 2090, 10)))
+  # years <- paste0('X', c(2015, seq(2020, 2090, 10)))
+  years <- colnames(data)[grepl('X', colnames(data))]
 
   # join weight to the data
   yield_impact_clean <- dplyr::bind_rows(
@@ -424,31 +425,37 @@ get_weighted_yield_impact <- function(data = NULL,
 #' Calculate agricultural productivity change based on yield impact multiplier and no impact yield
 #'
 #' @param data Default = NULL. output data from function yield_shock_projection, or similar format of data
-#' @param from_year Default = NULL. integer for 'from' year to calculate agricultural productivity change
-#' @param to_year Default = NULL. integer for 'to' year to calculate agricultural productivity change
+#' @param year_pairs Default = NULL. data frame for the paired years of from_year and to_year to calcualte ag productivity change
+#' @param gcam_timestep Default = 5. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)
 #'
 #' @keywords internal
 #' @export
 
 
 get_agprodchange <- function(data = NULL,
-                             from_year = NULL,
-                             to_year = NULL)
+                             # from_year = NULL,
+                             # to_year = NULL,
+                             year_pairs = NULL,
+                             gcam_timestep = 5)
 {
 
   year <- yield_multiplier <- AgProdChange_ni <- AgProdChange <- NULL
 
-  y1 <- paste0('X', from_year)
-  y2 <- paste0('X', to_year)
+  # Iterate over each year pair and calculate AgProdChange
+  for (i in seq_len(nrow(year_pairs))) {
+    y1 <- paste0('X', year_pairs$from_year[i])
+    y2 <- paste0('X', year_pairs$to_year[i])
 
-  df <- data %>%
-    dplyr::mutate(AgProdChange = ifelse(
-      year == y2,
-      ((yield_multiplier[year == y2] / yield_multiplier[year == y1])^(1/5)) * (1 + AgProdChange_ni[year == y2]) - 1,
-      AgProdChange)) %>%
-    dplyr::mutate(AgProdChange = ifelse(is.na(AgProdChange) & is.na(yield_multiplier), AgProdChange_ni, AgProdChange))
+    # Apply the transformation for the current year pair
+    data <- data %>%
+      dplyr::mutate(AgProdChange = ifelse(
+        year == y2,
+        ((yield_multiplier[year == y2] / yield_multiplier[year == y1])^(1 / gcam_timestep)) * (1 + AgProdChange_ni[year == y2]) - 1,
+        AgProdChange)) %>%
+      dplyr::mutate(AgProdChange = ifelse(is.na(AgProdChange) & is.na(yield_multiplier), AgProdChange_ni, AgProdChange))
+  }
 
-  return(df)
+  return(data)
 
 }
 
@@ -462,30 +469,33 @@ get_agprodchange <- function(data = NULL,
 #' Output agprodchange XML
 #'
 #' @param data Default = NULL. output data from function yield_shock_projection, or similar format of data
+#' @param gcamdata_dir Default = NULL. string for directory to the gcamdata folder within the specific GCAM version. The gcamdata need to be run with drake to have the CSV outputs beforehand.
 #' @param climate_model Default = 'gcm'. string for climate model name (e.g., 'CanESM5')
 #' @param climate_scenario Default = 'rcp'. string for climate scenario name (e.g., 'ssp245')
[…]
