+++ R/yield_impact.R
@@ -16,6 +16,8 @@
 #' @param bias_adj Default = 'ba'. string for the dataset used for climate data bias adjustment
 #' @param cfe Default = 'no-cfe'. string for whether the yield impact formula implemented CO2 fertilization effect
 #' @param gcam_version Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7
+#' @param gcam_timestep Default = 5. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)
+#' @param gcamdata_dir Default = NULL. string for directory to the gcamdata folder within the specific GCAM version. The gcamdata need to be run with drake to have the CSV outputs beforehand.
 #' @param use_default_coeff Default = FALSE. binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.
 #' @param base_year Default = NULL. integer for the base year (for GCAM)
 #' @param start_year Default = NULL. integer for the  start year of the projected data
@@ -42,6 +44,8 @@ yield_impact <- function(pr_hist_ncdf = NULL,
                          bias_adj = 'ba',
                          cfe = 'no-cfe',
                          gcam_version = 'gcam7',
+                         gcam_timestep = 5,
+                         gcamdata_dir = NULL,
                          use_default_coeff = FALSE,
                          base_year = NULL,
                          start_year = NULL,
@@ -119,6 +123,7 @@ yield_impact <- function(pr_hist_ncdf = NULL,
                                                       base_year = base_year,
                                                       start_year = start_year,
                                                       end_year = end_year,
+                                                      gcam_timestep = gcam_timestep,
                                                       smooth_window = smooth_window,
                                                       diagnostics = diagnostics,
                                                       output_dir = output_dir
@@ -132,6 +137,8 @@ yield_impact <- function(pr_hist_ncdf = NULL,
                                                   bias_adj = bias_adj,
                                                   cfe = cfe,
                                                   gcam_version = gcam_version,
+                                                  gcam_timestep = gcam_timestep,
+                                                  gcamdata_dir = gcamdata_dir,
                                                   diagnostics = diagnostics,
                                                   output_dir = output_dir)
 
