### 2024-11-25 #7393 @jrising
@cheginit Yes, I can do this. The next 6 weeks are challenging though. I would prefer to complete the work in early January. Is that in line with your 4-6 weeks, starting within 2 weeks, or would you need the initial review to be finalized before then?
