### 2024-11-26 #7393 @jrising
@cheginit I've previously reviewed for JOSE, so I understand that. But what is not clear to me from your statement is, are you asking me to finish the iteration in mid-January? I'm asking if it's alright for the initial review, prior to iteration, be ready for mid-January.
