### 2024-11-26 #7393 @cheginit
@jrising Thanks for being an active member of the JOSS community! To clarify, yes, it would be alright if you begin the review early January.
