### 2024-11-28 #7538 @editorialbot
[REVIEW]: gaia: An R package to estimate crop yield responses to temperature and precipitation
**Submitting author:** @mengqi-z (<a href="http://orcid.org/0000-0001-5385-2758">Mengqi Zhao</a>)
**Repository:** https://github.com/JGCRI/gaia
**Branch with paper.md** (empty if default branch): main
**Version:** v2.0.1
