### 2024-11-28 #7538 @cheginit
👋🏼 @mengqi-z, @klau506, and @jrising, this is the review thread for the paper. All of our communications will happen here from now on.

As a reviewer, the first step, as mentioned in the first comment of this issue, is to create a checklist for your review by entering

```@editorialbot generate my checklist```

The JOSS review is different from most other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#7538` so that a link is created to this thread (and I can keep an eye on what is happening). Please also feel free to comment and ask questions on this thread. In my experience, it is better to post comments/questions/suggestions as you come across them, instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. Please notify me if any of you require some more time. We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.

Please don't hesitate to ping me (`@cheginit`) if you have any questions/concerns.
