### 2024-12-09 JGCRI/gaia#4 @klau506
Manuscript suggestions
[review process #7538](https://github.com/openjournals/joss-reviews/issues/7538)

- L23: consider "powerful *tool*" instead of "powerful *model*" since `gaia` is a tool that embeds a model.
- L22-L24: you mention *economic models* and later *human-Earth system analysis*. Maybe it would be good to mention *integrated assessment models* instead of *economic models*?
- L29: consider rephrasing the sentence. E.g., "Agricultural markets are global, so yield shocks in one region can significantly affect others due to international trade".
- L33: from my point of view, *GCAM* is the acronym and is not part of the citation. Thus, it is better to write it as "... Global Change Analysis Model (GCAM) (citation)".
- L36: consider removing the comma *Most economic multisectoral...*
- L39: consider adding an *s*: *multiple cropS*
- Consider combining the sections Statement of Need and States of the Field into one bigger section, which indicates the necessity of `gaia` providing some reasons and the current state of the field.
- L65: consider writing the function names, such as `yield_impact`, inside backticks, as you do when mentioning the pkg.
- L72: daily *and* monthy...
- L84: consider mentioning that the fitted model is the one obtained in `yield_regression`.
- The *Functionality* section might benefit from a brief example on how to run the pkg, as done in a lot of  JOSS publications.
- Looking at the documentation, I understand that the outputs are by GCAM region, such as *South America Northern*, which are not countries. Consider the word *region* instead of *countries* in the whole manuscript and documentation. It will be more accurate.
- Bib: sometimes Waldhoff et al 2020 is cited as a hole, whereas sometimes only the "2020" year is clickable. Homogenize for consistency.
- Bib: Lizumi \& Ramankuty 2016 missing the capital letter of first author.
