### 2024-12-09 JGCRI/gaia#5 @klau506
Code suggestions
[review process #7538](https://github.com/openjournals/joss-reviews/issues/7538)

- R/crop_calendars.R: add `output_dir` as `@param`.
- R/crop_calendars.R: do or remove the *TODO* sentence in line 358.
- R/data_aggregation.R: the `utils` pkg is not required to be loaded specifically in the function's description since it is a default R pkg. Indeed, `data_aggregation` is the only function that calls the `@importFrom`, so consider removing it for the sake of consistency. In the `glob2rx` calls in the function, consider adding the `utils` pkg (`utils::glob2rx`) to all of them for the sake of consistency.
- R: consider using capital letter after the sentence *"Default = NULL"* repeatedly written in the functions' descriptions.
- R: consider adding the `@return` field in the functions' descriptions to clarify the functions' outputs.
- R: consider running the entire pkg through the `styler` pkg, with a style of your choice, to ensure clear and consistent code style.
