### 2024-12-09 JGCRI/gaia#6 @klau506
Documentation suggestions
[review process #7538](https://github.com/openjournals/joss-reviews/issues/7538)

- User Guide/Workflow: consider rephrasing the first paragraph. The sentences are very long and it's difficult to follow.
- Run gaia!/Example 2: when running this example, it stopped with the following Error message
```r
[1] "Estimated yield impacts data saved to: XXX/gaia_output/yield_impacts_annual/yield_canesm5_gcam-ref_wheat.csv"
[1] "Projected annual yield figure saved to: XXX/gaia_output/figures/annual_projected_climate_impacts_canesm5_gcam-ref_wheat_fit_lnyield_mmm_quad_noco2_nogdp.pdf"
[1] "Smoothed yield impacts saved to: XXX/gaia_output/yield_impacts_smooth/yield_canesm5_gcam-ref_wheat.csv"
[1] "Projected and smoothed yield figure saved to: XXX/gaia_output/figures/smooth_projected_climate_impacts_canesm5_gcam-ref_wheat_fit_lnyield_mmm_quad_noco2_nogdp.pdf"
[1] "Spatial map plot: canesm5 gcam-ref wheat 2090"
Error in `as_tibble()`:
! All columns in a tibble must be vectors.
x Column `geometry` is a `sfc_MULTIPOLYGON/sfc`
  object.
```
- Run gaia!/Example 3: consider removing the pkg name before the functions since `library(gaia)` was already called in the script (e.g., `crop_calendars` instead of `gaia::crop_calendars`)
- Run gaia!/Example 3: the sentence *"To view the country names associated with these IDs, simply type `country_id` in the R console after loading the `gaia` package."* does not work for me. The R console says `Error: object 'country_id' not found` after a successful run of the function and typing `country_id` in the R console.
- News tab: consider adding a brief description of the release.
- Contribute tab: the sentence *"Trivial changes to comments or documentation do not require creating a new issue."* is misleading. How should these trivial changes be reported then?
- Support tab: the links do not work. It says that the site is not available. It might be that the URL is wrongly introduced or a private/public repository issue.
