### 2024-12-09 #7538 @klau506
Dear @cheginit and @mengqi-z,

This is a highly valuable contribution with significant potential to be integrated directly into the GCAM integrated assessment model or employed as an external tool for further analyses. Overall, the manuscript is well-written and clear, though it would benefit from minor revisions and the addition of a brief example demonstrating how to execute the main package function.

The code is exceptionally well-organized, tidy, and thoroughly commented, which is commendable for promoting transparency and facilitating further contributions to the project. Additionally, the documentation is informative and well-structured, although I encountered a broken example. More specific feedback and suggestions for improvement are provided in the issues I have opened in the `gaia` repository.

I hope this is useful! I appreciate that the authors have done a lot of work, and I am happy to clarify/answer questions if useful.
