+++ R/yield_shock_projection.R
@@ -6,7 +6,7 @@
 #' @param use_default_coeff Default = FALSE. binary for using default regression coefficients. Set to TRUE will use the default coefficients instead of calculating coefficients from the historical climate data.
 #' @param climate_model Default = NULL. string for climate model (e.g., 'CanESM5')
 #' @param climate_scenario Default = NULL. string for climate scenario (e.g., 'ssp245')
-#' @param base_year Default = NULL. integer for the base year (for GCAM)
+#' @param base_year Default = 2015. integer for the base year (for GCAM)
 #' @param start_year Default = NULL. integer for the  start year of the data
 #' @param end_year Default = NULL. integer for the end year of the data
 #' @param gcam_timestep Default = 5. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)
@@ -18,7 +18,7 @@
 yield_shock_projection <- function(use_default_coeff = FALSE,
                                    climate_model = 'gcm',
                                    climate_scenario = 'rcp',
-                                   base_year = NULL,
+                                   base_year = 2015,
                                    start_year = NULL,
                                    end_year = NULL,
                                    gcam_timestep = 5,
