+++ R/pkg_utils.R
@@ -230,17 +230,81 @@ merge_data <- function( d1, d2, x1, x2 )
 
 
 # ------------------------------------------------------------------------------
+#' agprodchange_interp
+#'
+#' Check if the reference agricultural productivity has the same timestep as the user defined timestep
+#' If not, linearly interpolate the value
+#'
+#' @param data = NULL. data frame of the agprodchange
+#' @param gcam_timestep Default = 5. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)
+#' @keywords internal
+#' @export
+
+agprodchange_interp <- function(data = NULL,
+                                gcam_timestep = 5){
+
+  # check if the time step is the same with the APG no climate impact
+  # for example, the default GCAM is 5 year and reference APG from GCAM starts from 2020
+  # however, it the time step is 1 year, adding base year 2015 to the dataset will create a 5 year gap between 2015 and 2020
+  # In such case, we can linearly interpolate the values from the base year to the first future year
+  years_ref <- sort(as.integer(gsub('X', '', unique(data$year))))
+  ref_timestep_series <- years_ref[2:length(years_ref)] - years_ref[1:length(years_ref)-1]
+  ref_timestep <- unique(ref_timestep_series)
+
+  # find where the timestep series is difference from the gcam_timestep
+  ref_timestep_index <- which(ref_timestep_series != gcam_timestep)
+  ref_timestep_interp_start <- years_ref[ref_timestep_index]
+  ref_timestep_interp_end <- years_ref[ref_timestep_index + 1]
+
+  # check if the reference time step equals to user defined gcam_timestep
+
+  if(!any(grepl(gcam_timestep, ref_timestep)) | length(ref_timestep) != 1){
+
+    warning(paste0('The time step of the reference agricultural productivity change data is different from the user defined gcam_timestep = ',
+                   gcam_timestep))
+
+    print(paste0('Interpolating the reference agricultural productivity change to ',
+                 gcam_timestep, ' year time step between: ',
+                 ref_timestep_interp_start, ' to ', ref_timestep_interp_end))
+
+    data_interp <- data %>%
+      dplyr::mutate(year = as.numeric(gsub('X', '', year))) %>%
+      dplyr::group_by(region, AgSupplySector, AgSupplySubsector, AgProductionTechnology) %>%
+      tidyr::complete(year = tidyr::full_seq(year, gcam_timestep)) %>%
+      dplyr::mutate(AgProdChange_ni = ifelse(is.na(AgProdChange_ni),
+                                          approx(x = year[!is.na(AgProdChange_ni)],
+                                                 y = AgProdChange_ni[!is.na(AgProdChange_ni)],
+                                                 xout = year)$y,
+                                          AgProdChange_ni)) %>%
+      dplyr::ungroup() %>%
+      dplyr::mutate(year = paste0('X', year))
+
+  } else {
+
+    data_interp <- data
+
+  }
+
+
+  return(data_interp)
+}
+
+# ------------------------------------------------------------------------------
 #' agprodchange_ref
 #'
 #' Get the reference agricultural productivity change based GCAM version
 #'
 #' @param gcam_version Default = 'gcam7'. string for the GCAM version. Only support gcam6 and gcam7
+#' @param gcam_timestep Default = 5. integer for the time step of GCAM (Select either 1 or 5 years for GCAM use)
+#' @param base_year Default = 2015. integer for the base year (for GCAM)
 #' @param climate_scenario Default = NULL. string for climate scenario (e.g., 'ssp245')
[…]
