### 2024-12-20 63f50f8
Update JOSS paper
