### 2024-12-20 JGCRI/gaia#4 @mengqi-z
Hi @klau506,

Thank you for your comments! We have addressed them where appropriate and provided the following explanations for certain cases. Please refer to the commit 63f50f8cdfca7632f2fcdd9c1ece0e385432e420

> - L22-L24: you mention economic models and later human-Earth system analysis. Maybe it would be good to mention integrated assessment models instead of economic models?

We did not use the term "integrated assessment model" due to preferences expressed by our sponsor. Instead, we revised the text to use the term "multisectoral economic model".

> - Consider combining the sections Statement of Need and States of the Field into one bigger section, which indicates the necessity of gaia providing some reasons and the current state of the field.

Thank you for the suggestion. While both sections discuss the necessity of `gaia`, they emphasize different aspects. The Statement of Need section highlights the research gaps that `gaia` addresses, whereas States of the Field focuses on the current modeling landscape. To preserve these distinct focuses, we have decided to keep the sections separate.

> - The Functionality section might benefit from a brief example on how to run the pkg, as done in a lot of JOSS publications.

We appreciate this great suggestion. However, due to the JOSS word count limit (1000), we are unable to include a example because we are already over the word limit. Instead, we revised the sentence in the Functionality section to direct readers to our GitHub repository for detailed installation instructions and example:
"Detailed instructions on installing and using `gaia` with example dataset can be accessed at https://jgcri.github.io/gaia."

> - Looking at the documentation, I understand that the outputs are by GCAM region, such as South America Northern, which are not countries. Consider the word region instead of countries in the whole manuscript and documentation. It will be more accurate.

Thank you for pointing this out. `gaia`'s outputs are available at multiple spatial scales, as it supports both GCAM and non-GCAM models. For GCAM model, the outputs include agricultural productivity change at the region-basin intersections, while for other models, country-level outputs are available for those requiring country-specific yield shocks.

> - Bib: sometimes Waldhoff et al 2020 is cited as a hole, whereas sometimes only the "2020" year is clickable. Homogenize for consistency.

Thank you for noting this. This appears to be due to the default formatting in JOSS, where the hyperlink placement depends on the citation style. Specifically, "(Waldhoff et al., 2020)" links the entire citation, whereas "Waldhoff et al. (2020)" only links the year.

> - Bib: Lizumi & Ramankuty 2016 missing the capital letter of first author.

This is a formatting issue. The first author's name begins with an "I" (not an "L"), which can appear unclear due to the font style used in the JOSS paper.

Thanks again for your detailed comments on the paper!
