### 2024-12-20 f77aa25
Update function documentations and format code with styler
Update based on the suggestions from JOSS reviewer #5
