+++ R/data_aggregation.R
@@ -2,37 +2,35 @@
 #'
 #' Process multiple models to analyze historical weather variables and crop yields
 #'
-#' @param climate_hist_dir Default = NULL. string for path to the processed historical climate data folder
-#' @param climate_impact_dir Default = NULL. string for path to the processed future climate data folder using weighted_climate function
-#' @param climate_model Default = NULL. string for climate model (e.g., 'CanESM5')
-#' @param climate_scenario Default = NULL. string for climate scenario (e.g., 'ssp245')
-#' @param historical_periods Default = NULL. vector for years to subset from the historical climate data. If NULL, use the default climate data period
-#' @param co2_hist Default = NULL. data table for historical CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
-#' @param co2_proj Default = NULL. data table for projected CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
+#' @param climate_hist_dir Default = NULL. String for path to the processed historical climate data folder
+#' @param climate_impact_dir Default = NULL. String for path to the processed future climate data folder using weighted_climate function
+#' @param climate_model Default = NULL. String for climate model (e.g., 'CanESM5')
+#' @param climate_scenario Default = NULL. String for climate scenario (e.g., 'ssp245')
+#' @param historical_periods Default = NULL. Vector for years to subset from the historical climate data. If NULL, use the default climate data period
+#' @param co2_hist Default = NULL. Data table for historical CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
+#' @param co2_proj Default = NULL. Data table for projected CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
 #' @param output_dir Default = file.path(getwd(), 'output'). String for output directory
-#' @importFrom utils glob2rx
+#' @returns A list of historical and projected weather variables and crop data.
 #' @export
 
 data_aggregation <- function(climate_hist_dir = NULL,
                              climate_impact_dir = NULL,
-                             climate_model = 'gcm',
-                             climate_scenario = 'rcp',
+                             climate_model = "gcm",
+                             climate_scenario = "rcp",
                              historical_periods = NULL,
                              co2_hist = NULL,
                              co2_proj = NULL,
-                             output_dir = file.path(getwd(), 'output'))
-{
-
-  message('Starting Step: data_aggregation')
+                             output_dir = file.path(getwd(), "output")) {
+  message("Starting Step: data_aggregation")
 
   # read crop calendar output from crop_calendars function
   crop_cal <- gaia::input_data(
-    folder_path = file.path(output_dir, 'data_processed'),
-    input_file = 'crop_calendar.csv'
+    folder_path = file.path(output_dir, "data_processed"),
+    input_file = "crop_calendar.csv"
   )
 
   # merge fao yield and fao irrigation equip data
-  yield <- gaia::merge_data(fao_yield, fao_irr_equip, 'iso', 'year')
+  yield <- gaia::merge_data(fao_yield, fao_irr_equip, "iso", "year")
 
 
   # ----------------------------------------------------------------------------
@@ -40,67 +38,70 @@ data_aggregation <- function(climate_hist_dir = NULL,
   # ----------------------------------------------------------------------------
 
   # check the historical climate data folder
-  if(is.null(climate_hist_dir)){
-
[…]
