### 2024-12-20 1097b16
Fix the issue with plot spatial maps
issue #6
Import sf should solve the issue when joining a tibble into an sf object.
