+++ NAMESPACE
@@ -42,6 +42,7 @@ export(yield_impact)
 export(yield_regression)
 export(yield_shock_projection)
 export(z_estimate)
+import(sf)
 importFrom(data.table,":=")
 importFrom(magrittr,"%>%")
 importFrom(stats,approx)
