### 2024-12-21 4ac3d02
Make country_id external data
For issue #6

Make the data table country_id external so that users can access this information to check the country ID and name
