### 2024-12-21 5866925
Update CONTRIBUTING, NEWS, SUPPORT tabs
For issue #6
