+++ vignettes/vignette.Rmd
@@ -306,14 +306,14 @@ library(gaia)
 output_dir <- file.path(getwd(), 'gaia_output')
 
 # calculate weigted climate
-gaia::weighted_climate(pr_ncdf = pr_projection_file ,
-                       tas_ncdf = tas_projection_file ,
-                       timestep = 'monthly',
-                       climate_model = 'canesm5',
-                       climate_scenario = 'gcam-ref',
-                       time_periods = seq(2015, 2100, 1),
-                       output_dir = output_dir,
-                       name_append = NULL)
+weighted_climate(pr_ncdf = pr_projection_file ,
+                 tas_ncdf = tas_projection_file ,
+                 timestep = 'monthly',
+                 climate_model = 'canesm5',
+                 climate_scenario = 'gcam-ref',
+                 time_periods = seq(2015, 2100, 1),
+                 output_dir = output_dir,
+                 name_append = NULL)
 
 ```
 
@@ -349,7 +349,7 @@ The crop calendar outlines the planting and harvesting months for each predefine
 output_dir <- file.path(getwd(), 'gaia_output')
 
 # calculate crop calendars
-crop_cal <- gaia::crop_calendars(output_dir = output_dir)
+crop_cal <- crop_calendars(output_dir = output_dir)
 
 # print result
 crop_cal
@@ -395,21 +395,21 @@ The `data_aggregation` function creates two types of files in the output_dir/dat
 output_dir <- file.path(getwd(), 'gaia_output')
 
 # aggregate crop and climate information at the country level
-data_agg <- gaia::data_aggregation(climate_hist_dir = climate_hist_dir,
-                                   climate_impact_dir = climate_impact_dir,
-                                   climate_model = 'canesm5',
-                                   climate_scenario = 'gcam-ref',
-                                   output_dir = output_dir)
+data_agg <- data_aggregation(climate_hist_dir = climate_hist_dir,
+                             climate_impact_dir = climate_impact_dir,
+                             climate_model = 'canesm5',
+                             climate_scenario = 'gcam-ref',
+                             output_dir = output_dir)
 
 ```
 
 ```{r eval=T, echo=F, message=F, results='hide'}
 
-crop_projection <- gaia::input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
-                                    input_file = 'weather_canesm5_gcam-ref_soybean.csv')
+crop_projection <- input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
+                              input_file = 'weather_canesm5_gcam-ref_soybean.csv')
 
-crop_hist <- gaia::input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
-                              input_file = 'historic_vars_soybean.csv')
+crop_hist <- input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
+                        input_file = 'historic_vars_soybean.csv')
 
 ```
 
@@ -445,8 +445,8 @@ The `gaia::yield_regression` function performs empirical model fitting and gener
 output_dir <- file.path(getwd(), 'gaia_output')
 
 # perform empirical regression
-gaia::yield_regression(diagnostics = TRUE,
-                       output_dir = output_dir)
+yield_regression(diagnostics = TRUE,
+                 output_dir = output_dir)
 
 ```
 
@@ -491,15 +491,15 @@ Once the `gaia` model has completed fitting, the `gaia::yield_shock_projection` 
 output_dir <- file.path(getwd(), 'gaia_output')
 
 # calculate projected yield shocks
-out_yield_shock <- gaia::yield_shock_projection(use_default_coeff = FALSE,
-                                                climate_model = 'canesm5',
-                                                climate_scenario = 'gcam-ref',
-                                                base_year = 2015,
-                                                start_year = 2015,
[…]
