### 2024-12-21 c1fd35f
Update vignette.Rmd
