+++ vignettes/vignette.Rmd
@@ -405,11 +405,11 @@ data_agg <- data_aggregation(climate_hist_dir = climate_hist_dir,
 
 ```{r eval=T, echo=F, message=F, results='hide'}
 
-crop_projection <- input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
-                              input_file = 'weather_canesm5_gcam-ref_soybean.csv')
+crop_projection <- gaia::input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
+                                    input_file = 'weather_canesm5_gcam-ref_soybean.csv')
 
-crop_hist <- input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
-                        input_file = 'historic_vars_soybean.csv')
+crop_hist <- gaia::input_data(folder_path = file.path(getwd(), 'vignetteFigs'),
+                              input_file = 'historic_vars_soybean.csv')
 
 ```
 
