### 2024-12-21 JGCRI/gaia#6 @mengqi-z
Hi @klau506,

Thank you for catching those issues! I’ve addressed them and updated various parts of the documentation based on your feedback.

The error message mentioned in your second bullet point occurs because the `sf` package was not loaded. I’ve added `@importFrom sf` to the description, which should hopefully resolve the issue. If the fix doesn’t work as expected, please let me know. Thank you!
