### 2025-01-02 JGCRI/gaia#5 @klau506
Dear @mengqi-z ,

Thanks for your work! It looks great! However, the `@return` parameter has not been added to all functions. For example, it is still missing in the `weighted_climate` or `yield_regression` functions, among others. I would encourage to include it to all functions to facilitate the users' comprehension.
