### 2025-01-02 JGCRI/gaia#6 @klau506
Dear @mengqi-z ,

Thank you for your work! The examples now run seamlessly, the documentation is much more comprehensive, and all the links work perfectly 🎉
