### 2025-01-02 JGCRI/gaia#4 @klau506
Dear @mengqi-z ,

Thank you for your work. Some more minor comments are below:

- The term $CO_2$ is not written homogeneously. Sometimes it is written like _CO2_ and sometimes like $CO_2$. Homogenize for consistency.
- Maybe the _Functionality_ section would benefit from detailing the region-basin vs country level detail for GCAM and non-GCAM models since it is not straightforward.

Apart from these two points, I think the manuscript is very well written!!
