### 2025-01-02 78ff995
Update paper.md
For #4
- Update CO2 to be the consistent as CO~2~ through out the paper
- Improve Functionality section to specify the output application for GCAM and non-GCAM models.
