+++ paper/paper.md
@@ -28,7 +28,7 @@ bibliography: paper.bib
 
 # Summary
 
-`gaia` is an open-source R package designed to estimate crop yield shocks in response to annual weather variations and CO2 concentrations at the country scale for 12 major crops. This innovative tool streamlines the workflow from raw climate data processing to projections of annual shocks to crop yields at the country level, using the response surfaces developed and documented in @Waldhoff_2020, an empirical econometric model that leverages historical weather, CO~2~, and crop yield data for robust empirical fitting for 12 crops. `gaia` uses these response surfaces with bias-corrected, gridded monthly temperature and precipitation projections (e.g., from the Coupled Model Intercomparison Project Phase 6 (CMIP6) [@Oneil_2016] and Inter-Sectoral Impact Model Intercomparison Project (ISIMIP) [@Warszawski_2014]) to project shocks that can be applied to agricultural productivity changes at the country-level for use in multisectoral economic models. The historical and future projections use gridded, country-crop specific monthly growing season precipitation and temperature data, aggregated to the national level, weighted by cropland area derived from MIRCA [@Portmann_2010]. These annual, country, crop-specific yield shocks can be aggregated to different definitions of regions, crop commodities, and time periods, as needed by specific multisectoral economic models. `gaia` serves as a lightweight, powerful tool that can aid exploration of crop yields responses under a broad range of future climate projections, enhancing human-Earth system analysis capabilities.
+`gaia` is an open-source R package designed to estimate crop yield shocks in response to annual weather variations and CO~2~ concentrations at the country scale for 12 major crops. This innovative tool streamlines the workflow from raw climate data processing to projections of annual shocks to crop yields at the country level, using the response surfaces developed and documented in @Waldhoff_2020, an empirical econometric model that leverages historical weather, CO~2~, and crop yield data for robust empirical fitting for 12 crops. `gaia` uses these response surfaces with bias-corrected, gridded monthly temperature and precipitation projections (e.g., from the Coupled Model Intercomparison Project Phase 6 (CMIP6) [@Oneil_2016] and Inter-Sectoral Impact Model Intercomparison Project (ISIMIP) [@Warszawski_2014]) to project shocks that can be applied to agricultural productivity changes at the country-level for use in multisectoral economic models. The historical and future projections use gridded, country-crop specific monthly growing season precipitation and temperature data, aggregated to the national level, weighted by cropland area derived from MIRCA [@Portmann_2010]. These annual, country, crop-specific yield shocks can be aggregated to different definitions of regions, crop commodities, and time periods, as needed by specific multisectoral economic models. `gaia` serves as a lightweight, powerful tool that can aid exploration of crop yields responses under a broad range of future climate projections, enhancing human-Earth system analysis capabilities.
 
 
 # Statement of need
@@ -53,7 +53,7 @@ The primary functionality of `gaia` is encapsulated in the `yield_impact` wrappe
 2. `crop_calenders`: Generates crop planting months for each country and crop based on crop calendar data [@Sacks_2010].
 3. `data_aggregation`: Calculates crop growing seasons using climate variables processed by `weighted_climate` and crop calendars for both historical and projected periods. This function prepares climate and yield data for subsequent model fitting.
 4. `yield_regression`: Performs regression analysis fitted with historical annual crop yields, monthly growing season temperature and precipitation, CO~2~ concentrations, GDP per capita, and year. The default econometric model applied in `gaia` is from @Waldhoff_2020. User can specify alternative formulas that are consistent with the data processed in `data_aggregation`.
[…]
