### 2025-01-02 6992dbb
Update function documentations
For #5
- Add @returns description for functions with no return value.
