+++ man/plot_fit.Rd
@@ -17,6 +17,9 @@ plot_fit(
 
 \item{output_dir}{Default = file.path(getwd(), 'output'). String for output directory}
 }
+\value{
+No return value, called for the side effects of plotting fitted regression
+}
 \description{
 Figure of fitted regression estimates compared to FAO yields
 }
