### 2025-01-02 f71b533
Update paper.md
For #4
- Update the description for `gcam_agprodchange` to clarify its specific use to generate GCAM compatible outputs
