+++ paper/paper.md
@@ -54,7 +54,7 @@ The primary functionality of `gaia` is encapsulated in the `yield_impact` wrappe
 3. `data_aggregation`: Calculates crop growing seasons using climate variables processed by `weighted_climate` and crop calendars for both historical and projected periods. This function prepares climate and yield data for subsequent model fitting.
 4. `yield_regression`: Performs regression analysis fitted with historical annual crop yields, monthly growing season temperature and precipitation, CO~2~ concentrations, GDP per capita, and year. The default econometric model applied in `gaia` is from @Waldhoff_2020. User can specify alternative formulas that are consistent with the data processed in `data_aggregation`.
 5. `yield_shock_projection`: Projects country-level yield shocks for future climate scenarios using the fitted model obtained from `yield_regression` and temperature, precipitation, and CO~2~ projections from the climate scenario. The outputs can be integrated into non-GCAM modeling frameworks requiring country-level yield responses to climate.
-6. `gcam_agprodchange`: Remaps country-level yield shocks to GCAM-required spatial scales (i.e., region, basin, technology intersections), based on harvested areas, and aggregates crops to GCAM commodities. This function applies the projected shocks to GCAM scenario agricultural productivity growth rates (the unit used to project future yields in GCAM) and creates ready-to-use XML outputs for GCAM.
+6. `gcam_agprodchange`: Generates GCAM-compatible outputs for agricultural productivity changes by GCAM commodity at the region-basin level. The function remaps country-level yield shocks from `yield_shock_projection` to GCAM-required spatial scales (i.e., region, basin, technology intersections), using harvested areas and aggregates crops into GCAM commodities. This function applies these projected shocks to GCAM scenario-specific agricultural productivity growth rates (the metric used to project future yields in GCAM) and creates ready-to-use XML outputs for GCAM.
 
 
 ![The gaia workflow showing the functions and the corresponding outputs of modeling crop yield shocks to weather variables using an empirical econometric model. \label{fig:workflow}](workflow.jpg)
