### 2025-01-02 JGCRI/gaia#4 @mengqi-z
Dear @klau506,

Thank you for catching the CO2 inconsistency! I have updated the paper to address both of your comments.
