### 2025-01-02 JGCRI/gaia#5 @mengqi-z
Dear @klau506,

Thank you! I initially didn’t include `@returns` tag for few functions because they don’t return any value. However, you’re absolutely right that it’s still important to include the `@returns` tag to indicate additional information (e.g., functions are used to produce specific output files rather than returning a value). I’ve reviewed all the functions and ensured that each one now includes a `@returns` parameter.
