### 2025-01-14 JGCRI/gaia#10 @jrising
User Guide responses
https://github.com/openjournals/joss-reviews/issues/7538

These comments refer to the User Guide vignette (https://jgcri.github.io/gaia/articles/vignette.html).

 - What is the CMIP-ISIMIP format? I am not familiar with this data format: please provide a reference to a specification (also include this in example 2). Also, I wouldn't use raw CMIP data-- one needs downscaled data. Do you mean the protocol data produced for GGCMI?
 - [Sacks et al., (2020)](https://jgcri.github.io/doi/org/10.1111/j.1466-8238.2010.00551.x) link is broken. Also, that paper is from 2010, not 2020.
 - Sacks et al. reports day-of-year, not month-level crop calendars (as referred in the text). While the Waldhoff paper derives predictors from month-level data, much other work starts with daily data, so presumably you don't want to drop to month-level.
 - The user guide should include installation instructions.
 - It is very odd that you are translating data to ".txt format". This isn't even a format-- the extension just suggests that the file has some kind of text encoding.
 - Why do you recommend the use of WATCH climate observations? This is a very old dataset, based on an outdated ERA model, and you only need temperature and precipitation anyway.
 - It is odd that Example 1 under Run gaia! uses both Example 1 and Example 2 from the climate data section. If both are required, then these are not separate examples.
 - Run gaia! Example 2 refers to weighted climate data from Example Climate Data Example 2, but Example Climate Data Example 2 does not refer to weighted data.
 - It seems like Run gaia! Example 3 is just more information on the API and diagnostic outputs. It should be presented as such.
