### 2025-01-14 406251b
Fix the Slack et al., 2010 citation and link
For #10
