### 2025-01-15 JGCRI/gaia#10 @jrising
- https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/vignettes/vignette.Rmd#L126: This is identical to `output_dir <- 'gaia_output'`.
 - Most research that might want to use gaia will not be at the country level. Is this possible?
 - `weighted_climate`: What happens if the weather data does not cover all countries?
 - There is a printed error in the user guide under weighted_climate.
 - `crop_calendars`: You don't explain what this function does. I think it outputs the crop calendars-- possibly both as a return value and to the output directory?
 - You say that new crops can be added to the crop calendar under the existing format. I think you mean that one should add an additional column, filled with 0s, and then additional rows with 1 in that column for each country, and then fill in plan and harvest. It's not clear from the example of plant and harvest months are 1-based or 0-based. It's also not clear where this should be added-- maybe you mean that crop_calendars produces the default crop_calendar, but one wouldn't run it if one has a modified crop calendar with additional crops?
 - Table 3: Why does the output produced by `data_aggregation` have a separate yield for each month? What is `grow_season`? I can guess the others, but they should be defined. Why does Table 3 have temp and precip if the weather data is intended for table 4?
 - Table 4: What are the min and max calculated over? Is it over mean monthly temperature and total monthly precip?
 - `yield_regression`: Report the default yield regression used here.
 - Report the meaning of the output values for `yield_shock_projection`. I'm guessing it's the ratio of projected yields in each year to the projected yield in the base year.
