### 2025-01-15 JGCRI/gaia#10 @jrising
General comments:

 - Based on the User Guide, this strikes me as more as a repo repository than a tool. It is not clear that the provided functions can be used for a different project, and we are only given examples that are part of a single workstream.
 - Although the guide mentions in multiple places that inputs can be changed-- as long as they follow the same format-- the user guide does not particularly facilitate. The formats aren't provided, and it's often not clear how changes at one point will affect things later on. Someone just has to follow the single workstream and hope for the best.
 - The regression uses a log dependent variable, so a weighted average isn't theoretically justified. It's okay at higher resolutions, but at the country level, you end up with a log-sum-exp problem.
 - It should be clarified that projected climate data needs to be BCSD relative to the same observational dataset used in the regression.
 - Standard errors are reported in the regression, but I don't see how uncertainty is handled. Ignoring it wouldn't be consistent with the regression results.
