### 2025-01-15 85ad764
Add installation guide and clarify CMIP-ISIMIP data
For #10
