+++ paper/paper.md
@@ -49,7 +49,7 @@ These considerations may limit the usefulness of yield changes from process mode
 
 The primary functionality of `gaia` is encapsulated in the `yield_impact` wrapper function that streamlines the entire workflow shown in \autoref{fig:workflow}. The modular design also facilitates comprehensive diagnostic outputs, enhancing the tool’s utility for researchers and decision makers. Users can also execute individual functions to work through the main steps of the process. Detailed instructions on installing and using `gaia` with example dataset can be accessed at https://jgcri.github.io/gaia.
 
-1. `weighted_climate`: Processes CMIP-ISIMIP climate NetCDF data and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type. The function accepts both daily and monthly climate data that are consistent with the CMIP-ISIMIP NetCDF data format
+1. `weighted_climate`: Processes CMIP climate NetCDF data in accordance with the [ISIMIP simulation protocols](https://www.isimip.org/protocol/) and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type. The function accepts both daily or monthly climate data in NetCDF format that follows ISIMIP protocols.
 2. `crop_calenders`: Generates crop planting months for each country and crop based on crop calendar data [@Sacks_2010].
 3. `data_aggregation`: Calculates crop growing seasons using climate variables processed by `weighted_climate` and crop calendars for both historical and projected periods. This function prepares climate and yield data for subsequent model fitting.
 4. `yield_regression`: Performs regression analysis fitted with historical annual crop yields, monthly growing season temperature and precipitation, CO~2~ concentrations, GDP per capita, and year. The default econometric model applied in `gaia` is from @Waldhoff_2020. User can specify alternative formulas that are consistent with the data processed in `data_aggregation`.
