### 2025-01-16 JGCRI/gaia#10 @mengqi-z
@jrising - Thank you for your comments! We are working on addressing the issues you've highlighted and preparing responses to some questions you've raised. Thanks for the time and effort you put into reviewing our work!
