### 2025-01-16 JGCRI/gaia#10 @jrising
One more point: I was trying to follow the vignette, but the second example download (to https://zenodo.org/records/13976521/files/gaia_example_climate.zip) has timed out on me 5 times. I suggest you offer an option for people to download the files outside of that function, or reduce the size.
