### 2025-01-21 8876557
Update vignette
For #10

Update to clarify several instructions and languages in the user guides
