+++ vignettes/vignette.Rmd
@@ -64,7 +64,7 @@ font-style: italic;
 <!-------------------------->
 <!-------------------------->
 
-# Installtion
+# Installation
 
 <!-------------------------->
 <!-------------------------->
@@ -158,13 +158,17 @@ The primary functionality of `gaia` is encapsulated in the `yield_impact` wrappe
 <p align="center"><img src="vignetteFigs/divider.png"/></p>
 
 
-`gaia` requires global climate data from the Inter-Sectoral Impact Model Intercomparison Project ([ISIMIP](https://www.isimip.org/)) or data formatted similarly to ISIMIP data. Additionally, `gaia` supports climate data aggregated to a monthly time step. Due to the large size of daily time step climate data, we have provided an example of monthly aggregated climate data covering the period from 2015 to 2100.
+`gaia` requires global climate data from the Inter-Sectoral Impact Model Intercomparison Project ([ISIMIP](https://www.isimip.org/)) or data formatted according to ISIMIP protocols. Additionally, `gaia` supports climate data in both daily and monthly time step. Due to the large size of global climate data, we provide two types of example datasets tailored to different user needs.
+
+* [Example Data 1](#example-data-1-quick-start-dataset): This pre-processed climate dataset allows users to quickly run `gaia` without the need to process raw climate NetCDF data. It is an output from the `gaia::weighted_climate` function, which converts raw climate NetCDF data into cropland-weighted climate data by country. This dataset includes both historical and future pre-processed cropland-weighted precipitation and temperature by country.  If you want to quickly test `gaia` and view the outputs, this is the easiest way to get started.
+* [Example Data 2](#example-data-2-raw-climate-data): This climate dataset is designed for users who want to process raw climate NetCDF data using the `gaia::weighted_climate` function. It includes global monthly precipitation and temperature NetCDF files at a 0.5-degree resolution, covering the period from 2015 to 2100. Processing this data may take about an hour. If you intend to use your own climate datasets, this example will help you get set up.
+
 
 <br />
 
-## Example Data 1
+## Example Data 1: Quick Start Dataset
 
-Download the example data using the instructions below. This dataset includes both historical climate observations and future climate projections, already weighted by cropland areas and provided in .txt format. The cropland-area-weighted, country-level climate data contains monthly precipitation and temperature for 26 crop types, distinguishing between irrigated and rainfed areas.
+Download the example data 1 using the instructions below. This dataset includes both historical climate observations and future climate projections, weighted by cropland areas and formatted in a tabular structure (see [Table 1](#table1)), as required by `gaia`. The cropland-area-weighted, country-level climate data contains monthly precipitation and temperature for 26 crop types, distinguishing between irrigated and rainfed areas.
 
 
 ```{r eval=F}
@@ -173,7 +177,7 @@ Download the example data using the instructions below. This dataset includes bo
 library(gaia)
 
 # Path to the output folder. Change it to your desired location
-output_dir <- file.path(getwd(), 'gaia_output')
+output_dir <- 'gaia_output'
 
 # Cropland-weighted historical and future climate data
 data_dir <- gaia::get_example_data(
@@ -193,14 +197,14 @@ climate_impact_dir <- file.path(data_dir, 'canesm5')
 <br />
 
 <div class="alert alert-info">
[…]
