### 2025-01-23 e55dc52
Add example to add crop to crop calendar and clarify the outputs from each function in the tutorial
For #10
