+++ vignettes/vignette.Rmd
@@ -350,7 +350,7 @@ Explore the results in your output folder. For a detailed explanation of the out
 
 ## Example 3
 
-Example 1 and 2 demonstrates how to use the wrapper function `yield_impact` to streamline the entire workflows shown in [Figure 1](#figure1). In Example 3, we provide additional instructions on using each function described in the [Workflow](#workflow) section with the example data and explain the associated outputs and diagnostic plots.
+Example 1 and 2 demonstrates how to use the wrapper function `yield_impact` to streamline the entire workflows shown in [Figure 1](#figure1). In Example 3, we provide additional instructions on using each function described in the [Workflow](#workflow) section with the example data and explain the associated outputs and diagnostic plots if any.
 
 <br />
 
@@ -363,6 +363,16 @@ Example 1 and 2 demonstrates how to use the wrapper function `yield_impact` to s
 The example below uses the monthly precipitation and temperature projections from 2015 to 2100. To run this climate data processing, please provide the file paths for the precipitation and temperature NetCDF files using the `pr_ncdf` and `tas_ncdf` arguments, respectively, and adjust the other arguments accordingly to match the specifics of your climate data.
 
 Please note that this step may take up to an hour to complete.
+
+**Output of the function:** The function has no return values. It also writes the following output files to the `output_dir/weighted_climate` folder:
+
+* `[climate-model]_[climate-scenario]_month_precip_country_irc_[crop-number]_[start-year]_[end-year].csv`: This file is the irrigated cropland-area weighted precipitation for certain climate model, scenario, and crop.
+
+* `[climate-model]_[climate-scenario]_month_precip_country_rfd_[crop-number]_[start-year]_[end-year].csv`: This file is the rainfed cropland-area weighted precipitation for certain climate model, scenario, and crop.
+
+* `[climate-model]_[climate-scenario]_month_tmean_country_irc_[crop-number]_[start-year]_[end-year].csv`: This file is the irrigated cropland-area weighted temperature for certain climate model, scenario, and crop.
+
+* `[climate-model]_[climate-scenario]_month_tmean_country_rfd_[crop-number]_[start-year]_[end-year].csv`: This file is the rainfed cropland-area weighted temperature for certain climate model, scenario, and crop.
 
 
 ```{r eval=F, echo=T}
@@ -408,7 +418,9 @@ knitr::kable(input_climate[1:12],
 
 ---
 
-The crop calendar outlines the planting and harvesting months for each predefined crop across different countries. Users can modify the crop calendar file to add additional crops, following the established format. Below is an example of the crop calendar structure. You can also access the crop calendar output at `output_dir/data_processed/crop_calendar.csv`.
+The `crop_calendars` function calculate crop calendars based on the SAGE global crop planting and harvesting dates data ([Sacks et al., (2010)](https://doi.org/10.1111/j.1466-8238.2010.00551.x)).  The crop calendar outlines the planting and harvesting months for each predefined crop across all global countries. To run `crop_calendars` following the code below.
+
+**Outputs of the function:** The function returns a data frame of crop calendars. It also writes the output in `output_dir/data_processed/crop_calendar.csv`  
 
 ```{r eval=F, echo=T}
 
@@ -422,6 +434,9 @@ crop_cal <- crop_calendars(output_dir = output_dir)
 crop_cal
 
 ```
+
+
[…]
