### 2025-01-27 JGCRI/gaia#11 @jrising
Code comments
https://github.com/openjournals/joss-reviews/issues/7538

Comments on code in `R`:

 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/crop_calendars.R#L26: Day-of-year divided by 30 isn't a great approximation for month, unless you're working with 360-day calendar data.
 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/crop_calendars.R#L47: This would be less error-prone if encoded in a data file.
 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/data_aggregation.R#L29: If you intend for users to modify this, provide them a way to use their own filename.
 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/data_aggregation.R#L51: Why is only the RFC weighted data used?
 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/weighted_climate.R#L98: This assumes a 1-to-1 mapping between region and grid cell. At the country level, this is usually fine, but would not work at higher resolutions.
 - https://github.com/JGCRI/gaia/blob/f4a0e78e01103869b235825c49f5a7f70639012d/R/yield_impacts_functions.R#L439: The SE should also be clustered, since there will be correlation between error terms in the each country.
