### 2025-01-27 6e0153a
Update data_aggregation in the user guide
For #10

Better clarify the differences between two output tables and add explanations of the variables in the outputs
