### 2025-01-28 af21a88
Update vignette to include the formular for the regression model
For #10
