### 2025-01-28 b8b2ad5
Update vignette to explain the meaning of yield shocks
For #10
