### 2025-01-28 f83561f
Update vignette to add more climate data related information
For #10
