+++ vignettes/vignette.Rmd
@@ -126,7 +126,7 @@ brew install gdal
 The primary functionality of `gaia` is encapsulated in the `yield_impact` wrapper function, which executes the entire workflow from climate data processing to yield shock estimation. Users can also execute individual functions to work through the main steps of the process ([Figure 1](#figure1)).
 
 
-1. [`weighted_climate`](#weighted_climate): Processes CMIP climate NetCDF data in accordance with the [ISIMIP simulation protocols](https://www.isimip.org/protocol "Here we specifically require the bias-adjusted and statistically downscaled CMIP climate data following ISIMIP's protocols") and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type. The function accepts both daily or monthly climate data in NetCDF format that follows ISIMIP protocols.
+1. [`weighted_climate`](#weighted_climate): Processes CMIP climate NetCDF data in accordance with the [ISIMIP simulation protocols](https://www.isimip.org/protocol/preparing-simulation-files/ "Here we specifically require the bias-adjusted and statistically downscaled CMIP climate data following ISIMIP's protocols") and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type. The function accepts both daily or monthly climate data in NetCDF format that follows ISIMIP protocols.
 
 2. [`crop_calendars`](#crop_calendars): Generates crop planting months for each country and crop based on crop calendar data [Sacks et al., (2010)](https://doi.org/10.1111/j.1466-8238.2010.00551.x).
 
@@ -170,6 +170,11 @@ The primary functionality of `gaia` is encapsulated in the `yield_impact` wrappe
 
 Download the example data 1 using the instructions below. This dataset includes both historical climate observations and future climate projections, weighted by cropland areas and formatted in a tabular structure (see [Table 1](#table1)), as required by `gaia`. The cropland-area-weighted, country-level climate data contains monthly precipitation and temperature for 26 crop types, distinguishing between irrigated and rainfed areas.
 
+There are two options to download the data:
+
+* Option 1: Please use this [download url](https://zenodo.org/records/13976521/files/weighted_climate.zip?download=1) and set the `data_dir` to the directory of the downloaded folder;
+* Option 2: Use `gaia::get_example_data` function following the code chunk below.
+
 
 ```{r eval=F}
 
@@ -184,7 +189,12 @@ data_dir <- gaia::get_example_data(
   download_url = 'https://zenodo.org/records/13976521/files/weighted_climate.zip?download=1',
   data_dir = output_dir
 )
+```
 
+
+Then, set the paths to historical and future climate data folder.
+
+```{r eval=F}
 # Path to the folder that holds cropland-area-weighted precipitation and temperature TXT files
 # historical climate observation
 climate_hist_dir <- file.path(data_dir, 'climate_hist')
@@ -206,6 +216,10 @@ climate_impact_dir <- file.path(data_dir, 'canesm5')
 
 Download the example data 2 using the instructions below. This dataset includes future global monthly precipitation and temperature from 2015 to 2100 at 0.5-degree resolution. To use a different climate model or scenario, you can provide your own climate data in NetCDF format. Note that `gaia` adheres to the ISIMIP climate data format as its standard, so your NetCDF files should be formatted accordingly if the data is not directly downloaded from ISIMIP.
 
+There are two options to download the data:
+
+* Option 1: Please use this [download url](https://zenodo.org/records/13976521/files/gaia_example_climate.zip?download=1) and set the `data_dir` to the directory of the downloaded folder;
+* Option 2: Use `gaia::get_example_data` function following the code chunk below.
 
 ```{r eval=F}
 
@@ -221,6 +235,11 @@ data_dir <- gaia::get_example_data(
   data_dir = output_dir
 )
 
+```
+
[…]
