### 2025-02-03 JGCRI/gaia#13 @jrising
Other repository comments
https://github.com/openjournals/joss-reviews/issues/7538

Other files in the repo:
 - `data-raw/DATASET.R` refers to files that a user wouldn't have. It seems helpful to include it, but tell users that it is not meant to be run, but where to get the underlying data.
 - `inst/extras/devTests.R` also would not be possible to run by users. Provide better comments at the top of the file.
 - The autogenerated files in `man` do not have useful descriptions. These should be understandable without users resorting to the vignette.
 - The unit tests only capture the vignette process, rather than considering multiple possible uses for each function.

JOSS paper comments:
 - CMIP6 generally does not provide downscaled, bias-corrected data. You should be more specific.
 - Also, the Zenodo archive mentions the use of STITCHES to prepare the climate data, but this is not discussed in the user guide.
 - I believe the section should be titled "State of the Field" not "States..."

Other points from the JOSS checklist:
 - I remain concerned that this is essentially still a reproduction repository, since it is not clear that it can be used in cases other than reproducing the results of the Waldhoff paper.
 - The repo includes several datasets, but I do not see documentation of those datasets.
 - I look forward to trying to reproduce the vignette results when I can download the associated data.
 - Why is RStudio considered a necessary installation step?
 - The contributing link in the README is broken.
