### 2025-02-14 01a0525
Add functionality to allow user to input crop calendar file
For #11 #10

- flexibility to use user provided crop calendar
- flexibility to select crops
- updated to the latest FAO data
- updated with FAO crop to MIRCA2000 crop mapping
- updated default MIRCA2000 crops to GCAM commodity mapping
- make cleaning sage data its own function
