+++ R/yield_impacts_functions.R
@@ -7,43 +7,89 @@
 #' Function to clean FAO yield data
 #'
 #' @param fao_yield Default = NULL. Data frame for the fao yield table
+#' @param fao_to_mirca Default = NULL. Data frame for the fao to mirca crop mapping
 #' @returns A data frame of formatted FAO yield data
 #' @keywords internal
 #' @export
 
-clean_yield <- function(fao_yield = NULL) {
+clean_yield <- function(fao_yield = NULL,
+                        fao_to_mirca = NULL) {
   AreaName <- crop <- NULL
 
-  d <- subset(fao_yield, select = c("AreaName", "ElementName", "ItemName", "Year", "Value"))
-  d <- subset(d, AreaName != "")
-  d <- subset(d, AreaName != "China, mainland")
-  d <- gaia::colname_replace(d, "AreaName", "country_name")
-  d <- gaia::colname_replace(d, "ElementName", "var")
-  d <- gaia::colname_replace(d, "ItemName", "crop")
-  d <- gaia::colname_replace(d, "Year", "year")
-  d <- gaia::colname_replace(d, "Value", "value")
-  d$crop <- tolower(d$crop)
-  d$var <- gsub("Area harvested", "area_harvest", d$var)
-  d$var <- gsub("Yield", "yield", d$var)
-  d <- subset(d, crop != "cottonseed") # Cotton seed has no data, FAO yields are for seed cotton
-  d$crop <- gsub("rice, paddy", "rice", d$crop)
-  d$crop <- gsub("soybeans", "soybean", d$crop)
-  d$crop <- gsub("sugar cane", "sugarcane", d$crop)
-  d$crop <- gsub("sugar beet", "sugarbeet", d$crop)
-  d$crop <- gsub("seed cotton", "cotton", d$crop)
-  d$crop <- gsub("sunflower seed", "sunflower", d$crop)
-  d$crop <- gsub(" ", "_", d$crop)
-  d <- data.table::dcast(d, country_name + crop + year ~ var, value.var = "value")
-  d <- merge(d, mapping_gcam_iso, by = "country_name", all.x = TRUE)
-  d <- gaia::iso_replace(d)
-  #   For these countries, root_tuber == cassava; others root_tuber == potato
-  d$crop <- ifelse(d$crop == "potatoes", "root_tuber", d$crop)
-  #   d$crop <- ifelse( ( ( d$iso == "ben" | d$iso ==  "cmr" | d$iso == "caf" | d$iso == "cog" | d$iso == "cod" | d$iso == "civ" | d$iso == "gab" | d$iso == "gha" |
-  #                           d$iso == "gin" | d$iso == "lbr" | d$iso == "nga" | d$iso == "sle" | d$iso == "som" | d$iso == "tgo" ) & d$crop == "cassava" ), "root_tuber", d$crop )
-  #   d$crop <- ifelse( ( ( d$iso == "ben" | d$iso ==  "cmr" | d$iso == "caf" | d$iso == "cog" | d$iso == "cod" | d$iso == "civ" | d$iso == "gab" | d$iso == "gha" |
-  #                           d$iso == "gin" | d$iso == "lbr" | d$iso == "nga" | d$iso == "sle" | d$iso == "som" | d$iso == "tgo" ) & d$crop == "potatoes" ), "potatoes", d$crop )
-  d <- subset(d, select = c("iso", "crop", "year", "area_harvest", "yield"))
-  return(d)
+  # clean up FAOSTAT yield and harvest area data and left join the iso code
+  df <- fao_yield %>%
+    dplyr::filter(`Item Code` %in% unique(fao_to_mirca$fao_crop_id),
+                  Element %in% c('Area harvested', 'Yield'),
+                  # Exclude entire China because FAOSTAT distinguishes China mainland, Taiwan, and Hongkong
[…]
