+++ R/yield_regression.R
@@ -4,12 +4,14 @@
 #' using average growing season temperature and precipitation, max and min months
 #'
 #' @param formula Default = NULL. String for regression formula
+#' @param crop_select Default = NULL. Vector of strings for the selected crops from our database. If NULL, the default crops will be used in the crop calendar: c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat"). The additional crops available for selection from our crop calendar database are: "barley", "groundnuts", "millet", "pulses", "rape_seed", "rye"
 #' @param diagnostics Default = TRUE. Logical for performing diagnostic plot
 #' @param output_dir Default = file.path(getwd(), 'output'). String for output directory
 #' @returns No return value, called for the side effects of processing and writing output files
 #' @export
 
 yield_regression <- function(formula = NULL,
+                             crop_select = NULL,
                              diagnostics = TRUE,
                              output_dir = file.path(getwd(), "output")) {
   # Decided not to use irrigation equipped land because data is so spotty
@@ -24,8 +26,12 @@ yield_regression <- function(formula = NULL,
   message(paste0("Starting step: yield_regression"))
   message(paste0("Using formula: ", formula))
 
+  if(is.null(crop_select)){
+    crop_select <- gaia::verify_crop(crop_select = crop_select)$crop_mirca
+  }
+
   # conduct regression for each crop
-  for (crop_i in mapping_mirca_sage$crop_mirca) {
+  for (crop_i in crop_select) {
     # read in the historical crop data and weather data
     d_crop <- gaia::input_data(
       folder_path = file.path(output_dir, "data_processed"),
