+++ R/crop_calendars.R
@@ -3,45 +3,173 @@
 #' Generate planting months for each country
 #' Data from SAGE
 #'
+#' @param crop_calendar_file Default = NULL. String for the path of the crop calendar file. If crop_cal is provided, crop_select will be set to crops in crop calendar. User provided crop_calendar_file can include any crops MIRCA2000 crops: "wheat", "maize", "rice", "barley", "rye", "millet", "sorghum", "soybean", "sunflower", "root_tuber", "cassava", "sugarcane", "sugarbeet", "oil_palm", "rape_seed", "groundnuts", "pulses", "citrus", "date_palm", "grapes", "cotton", "cocoa", "coffee", "others_perennial", "fodder_grasses", "others_annual"
+#' @param crop_select Default = NULL. Vector of strings for the selected crops from our database. If NULL, the default crops will be used in the crop calendar: c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat"). The additional crops available for selection from our crop calendar database are: "barley", "groundnuts", "millet", "pulses", "rape_seed", "rye"
 #' @param output_dir Default = file.path(getwd(), 'output'). String for output directory
 #' @returns A data frame of crop calendars
 #' @export
 
-crop_calendars <- function(output_dir = file.path(getwd(), "output")) {
+crop_calendars <- function(crop_calendar_file = NULL,
+                           crop_select = NULL,
+                           output_dir = file.path(getwd(), "output")) {
   iso <- crop <- NULL
 
   message("Starting Step: crop_calendars")
 
-  # load SAGE data
-  d <- sage
+  # verify the crops provided or selected are within the MIRCA crops
+  crops <- gaia::verify_crop(
+    crop_calendar_file = crop_calendar_file,
+    crop_select = crop_select)
 
-  # format, subset data and replace location name
-  d$Location <- ifelse(d$Location == "Georgia" & d$Source == "USDA UPHD", "Georgia_USA", d$Location)
-  d <- subset(d, select = c("Data.ID", "Location", "Crop", "Qualifier", "Plant.start", "Plant.end", "Harvest.start", "Harvest.end"))
-  d <- gaia::colname_replace(d, "Location", "country_name")
-  d <- gaia::colname_replace(d, "Crop", "crop")
-  d$crop <- tolower(d$crop)
+  # use user data if provided
+  if(!is.null(crop_calendar_file)){
 
-  # Average plant and harvest month
-  d$plant <- ifelse(d$Plant.start < d$Plant.end, ((d$Plant.start + d$Plant.end) / 2 / 30), ((d$Plant.start + d$Plant.end + 365) / 2 / 30))
-  d$plant <- ceiling(ifelse(d$plant > 12, d$plant - 12, d$plant))
-  d$plant <- ifelse(d$crop == "sugarcane", 1, d$plant) # Sugarcane is a multiyear crop, will assume that 12 months preceding annual value equals growing season.
-  d$harvest <- ifelse(d$Harvest.start < d$Harvest.end, ((d$Harvest.start + d$Harvest.end) / 2) / 30, ((d$Harvest.start + d$Harvest.end + 365) / 2) / 30)
-  d$harvest <- ceiling(ifelse(d$harvest > 12, d$harvest - 12, d$harvest))
[…]
