+++ R/crop_calendars.R
@@ -61,7 +61,8 @@ crop_calendars <- function(crop_calendar_file = NULL,
     }
 
     # clean up crop calendar
-    d <- gaia::clean_sage(d = d)
+    d <- gaia::clean_sage(d = d,
+                          crop_select = crops$crop_mirca)
 
   }
 
@@ -93,8 +94,9 @@ crop_calendars <- function(crop_calendar_file = NULL,
 #' @export
 
 verify_crop <- function(crop_calendar_file = NULL,
-                        crop_select = c("cassava", "cotton", "maize", "rice", "root_tuber", "sorghum", "soybean", "sugarbeet", "sugarcane", "sunflower", "wheat")){
+                        crop_select = NULL){
 
+  crop_cal <- NULL
 
   # Step 1: Finalize crop_select
   # read the user provided crop calendar
@@ -164,12 +166,14 @@ verify_crop <- function(crop_calendar_file = NULL,
 #' Clean up the sage crop calendar by reassigning the iso name and
 #' make sure to keep one set of the plant and harvest for each country and crop
 #'
-#' @param d Default = NULL. data frame for crop calendar
+#' @param d Default = NULL. Data frame for crop calendar
+#' @param crop_select Default = NULL. Vector of strings for selected crops
 #' @returns A data frame of clean crop calendar
 #' @keywords internal
 #' @export
 
-clean_sage <- function(d = NULL){
+clean_sage <- function(d = NULL,
+                       crop_select = NULL){
 
   # ----------------------------------------------------------------------------
   # Deal with exceptions (countries by states or regions)
@@ -398,12 +402,10 @@ clean_sage <- function(d = NULL){
   d$soybean <- ifelse(d$country_name == "Argentina" & d$crop == "soybeans" & d$Qualifier == "", 0, d$soybean)
 
   # Somalia (two seasons, use the one with no qualifier code)
-  d$cotton <- ifelse(d$country_name == "Somalia" & d$crop == "cotton" & d$Qualifier == 2, 0, d$cotton)
-  d$millet <- ifelse(d$country_name == "Somalia" & d$crop == "millet" & d$Qualifier == 2, 0, d$millet) #MZ
+  d$cotton <- ifelse(d$country_name == "Somalia" & d$crop == "cotton" & d$Data.ID == 1741, 0, d$cotton)
 
   # Kenya (two obs, nearly the same)
   d$cotton <- ifelse(d$country_name == "Kenya" & d$crop == "cotton" & d$Data.ID == 1711, 0, d$cotton)
-  d$millet <- ifelse(d$country_name == "Kenya" & d$crop == "millet" & d$Qualifier == 2, 0, d$millet) #MZ
 
   # SORGHUM CORRECTIONS
   # Dominican Republic (iso ID -- dom) listed thrice for sorghum
@@ -441,68 +443,85 @@ clean_sage <- function(d = NULL){
   d$iso <- ifelse((d$country_name == "North Korea (central bowl)" & d$crop == "soybeans"), "prk", d$iso)
   d$soybean <- ifelse(d$country_name == "North Korea (central bowl)" & d$crop == "soybeans", 1, d$soybean)
 
+  # Millet Corrections (MZ)
+  if('millet' %in% crop_select){
[…]
