### 2025-02-15 eccd3c2
Update  crop calendar instruction
