+++ vignettes/vignette.Rmd
@@ -448,7 +448,7 @@ knitr::kable(input_climate[1:12],
 
 ---
 
-The `crop_calendars` function calculates crop calendars based on the SAGE global crop planting and harvesting dates data ([Sacks et al., 2010](https://doi.org/10.1111/j.1466-8238.2010.00551.x)).  The crop calendar outlines the planting and harvesting months for each predefined crop across all global countries. The code chunk below runs `crop_calendars`.
+The `crop_calendars` function calculates crop calendars based on the SAGE global crop planting and harvesting dates data ([Sacks et al., 2010](https://doi.org/10.1111/j.1466-8238.2010.00551.x)).  The crop calendar outlines the planting and harvesting months for each predefined crop across all global countries. The code chunk below runs `crop_calendars` for the default crops: cassava, cotton, maize, rice, root_tuber, sorghum, soybean, sugarbeet, sugarcane, sunflower, wheat.
 
 
 ```{r eval=F, echo=T}
@@ -486,29 +486,42 @@ knitr::kable(crop_cal[1:10],
   footnote(general = 'This only shows the first 10 lines of the example data.')
 ```
 
+Beside default crops, users can select additional available crops (barley, groundnuts, millet, pulses, rape_seed, and rye) using `crop_select` argument. The crops that are available for selection in `gaia` are based on the crop availability from both MIRCA2000 and SAGE crop types. Note that FAOSTAT crops are categorized into MIRCA2000 crop category. In summary, `gaia` uses MIRCA2000 crop category as standard crop types. The example below selects barley and millet in addition to the default ones.
 
-Users can customize the crop calendar by adding new crops to the file, ensuring they follow the established format. The code snippet below shows an example on how to add a crop to the crop calendar. Once you've updated the crop calendar, be sure to save the modified file by overwriting the existing crop calendar CSV in your output folder. This ensures the changes will be applied in the subsequent steps.
+```{r eval=F, echo=T}
+
+# calculate crop calendars
+crop_cal <- crop_calendars(crop_select = c("barley", "millet"),
+                           output_dir = output_dir)
+
+# print result
+crop_cal
+
+```
+
+Due to limited crops from SAGE, some crops that are available in MIRCA2000 are not included in the SAGE database, including oil_palm, citrus, date_palm, grapes, cocoa, coffee, others_perennial, fodder_grasses, and others_annual. However, if users have crop planting and harvesting month data for any of these crops for certain countries, providing the crop calendar file is another way to include new crops. Of course, it would be the best if users can provide the planting and harvesting months of a crop for all countries that grow such crop for better results from empirical model fitting. Simply follow the established crop calendar format similar to [Table 2](#table2). The code snippet below shows an example on how to add a crop to the crop calendar. The name of the new crops need to follow the format of the crops listed above (e.g., use oil_palm with underscore instead of oil palm).
 
 ```{r eval=T, echo=T, message=F, results='hide'}
 
-# adding a new crop: yams. Construct the structure of the data with yams
+# adding a new crop: oil_palm. The crop name should follow the crop names listed above.
+# Construct the structure of the data with oil_palm
 crop_add <- expand.grid(iso = c('cog', 'gha', 'lbr'),
-                        crops = c(names(crop_cal)[2:(ncol(crop_cal) - 2)], 'yams')) %>%
-  dplyr::mutate(value = ifelse(crops == 'yams', 1, 0)) %>% 
+                        crops = c(names(crop_cal)[2:(ncol(crop_cal) - 2)], 'oil_palm')) %>%
[…]
