### 2025-02-17 efbe315
Update crop_calendars instruction in the vignette
For #10
Added more instruction on adding crops to crop calendar
