### 2025-02-18 b5c5036
Update example data 2 to smaller size for easier test
For #10
