+++ vignettes/vignette.R
@@ -2,6 +2,9 @@
 knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
 
 ## ----eval=F-------------------------------------------------------------------
+#  data_dir <- 'path/to/downloaded/folder'
+
+## ----eval=F-------------------------------------------------------------------
 #  
 #  # load gaia
 #  library(gaia)
@@ -11,7 +14,7 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #  
 #  # Cropland-weighted historical and future climate data
 #  data_dir <- gaia::get_example_data(
-#    download_url = 'https://zenodo.org/records/13976521/files/weighted_climate.zip?download=1',
+#    download_url = 'https://zenodo.org/records/14888816/files/weighted_climate.zip?download=1',
 #    data_dir = output_dir
 #  )
 
@@ -24,6 +27,9 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #  
 
 ## ----eval=F-------------------------------------------------------------------
+#  data_dir <- 'path/to/downloaded/folder'
+
+## ----eval=F-------------------------------------------------------------------
 #  
 #  # load gaia
 #  library(gaia)
@@ -33,7 +39,7 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #  
 #  # Future Climate Data
 #  data_dir <- gaia::get_example_data(
-#    download_url = 'https://zenodo.org/records/13976521/files/gaia_example_climate.zip?download=1',
+#    download_url = 'https://zenodo.org/records/14888816/files/gaia_example_climate.zip?download=1',
 #    data_dir = output_dir
 #  )
 #  
@@ -42,8 +48,8 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #  # Path to the precipitation and temperature NetCDF files
 #  # NOTE: Each variable can have more than one file
 #  # projected climate data
-#  pr_projection_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_gcam-ref_2015_2100.nc')
-#  tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_gcam-ref_2015_2100.nc')
+#  pr_projection_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_ssp245_2015_2030.nc')
+#  tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_ssp245_2015_2030.nc')
 #  
 
 ## ----eval=F, echo=T-----------------------------------------------------------
@@ -65,7 +71,7 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #    climate_hist_dir = climate_hist_dir,    # path to the folder that holds cropland weighted historical climate observations
 #    historical_periods = c(1960:2001),      # vector of historical years selected for fitting
 #    climate_model = 'canesm5',              # label of climate model name
-#    climate_scenario = 'gcam-ref',          # label of climate scenario name
+#    climate_scenario = 'ssp245',            # label of climate scenario name
 #    member = 'r1i1p1f1',                    # label of ensemble member name
 #    bias_adj = 'w5e5',                      # label of climate data for bias adjustment for the global climate model (GCM)
 #    cfe = 'no-cfe',                         # label of CO2 fertilization effect in the formula (default is no CFE)
@@ -73,10 +79,11 @@ knitr::opts_chunk$set(warning = FALSE, message = FALSE)
 #    use_default_coeff = FALSE,              # set to TRUE when there is no historical climate data available
 #    base_year = 2015                        # GCAM base year
 #    start_year = 2015,                      # start year of the projected climate data
[…]
