+++ tests/testthat/test-gaia.R
@@ -7,7 +7,7 @@ testthat::skip_on_travis()
 output_dir_i <- file.path(getwd(), 'output')
 
 data_dir_i <- gaia::get_example_data(
-  download_url = 'https://zenodo.org/records/13976521/files/weighted_climate.zip?download=1',
+  download_url = 'https://zenodo.org/records/14888816/files/weighted_climate.zip?download=1',
   data_dir = output_dir_i)
 
 
