### 2025-02-18 1efd5f4
Remove devTest.R
For #13

dev test script is not needed to be included in the package - it was used for developer's reference
