+++ tests/testthat/helper.R
@@ -10,7 +10,7 @@ output_dir_i <- file.path(getwd(), 'output')
 
 # setup variables
 climate_model_i <- 'canesm5'
-climate_scenario_i <- 'gcam-ref'
+climate_scenario_i <- 'ssp245'
 member_i = 'r1i1p1f1'
 bias_adj_i = 'w5e5'
 
@@ -27,6 +27,21 @@ smooth_window_i = 20
 diagnostics_i <- T
 use_default_coeff_i <- F
 
+# -------------------------------
+# Test Data
+# -------------------------------
+set.seed(7)
+
+# CO2
+co2_hist_i <- data.table::data.table(
+  year = 1959:2015,
+  co2_conc = seq(300, 400, length.out = length(1959:2015)) + rnorm(length(1959:2015), mean = 0, sd = 5)
+)
+
+co2_proj_i <- data.table::data.table(
+  year = 2015:2100,
+  co2_conc = 400 + (2015:2100-2000)^1.3 + rnorm(length(2015:2100), mean = 0, sd = 5)
+)
 
 # -------------------------------
 # Functions
@@ -50,6 +65,8 @@ run_crop_calendars <- function(crop_calendar_file = NULL,
 run_data_aggregation <- function(data_dir = NULL,
                                  climate_model = climate_model_i,
                                  climate_scenario = climate_scenario_i,
+                                 co2_hist = co2_hist_i,
+                                 co2_proj = co2_proj_i,
                                  output_dir = output_dir_i){
 
   climate_hist_dir_i <- file.path(data_dir, 'climate_hist')
@@ -59,6 +76,8 @@ run_data_aggregation <- function(data_dir = NULL,
                                    climate_impact_dir = climate_impact_dir_i,
                                    climate_model = climate_model,
                                    climate_scenario = climate_scenario,
+                                   co2_hist = co2_hist,
+                                   co2_proj = co2_proj,
                                    output_dir = output_dir)
 
   return(output)
