### 2025-02-18 c2dca32
Fix equation expression
