### 2025-02-19 5d9ac30
Update weighted_climate test
#13
- fixed the warnings from tidyr regarding using all_of() within across function
