### 2025-02-19 JGCRI/gaia#10 @mengqi-z
Dear @jrising,

Thank you so much for your detailed review of `gaia`. Your feedback has been so helpful for us in improving its functionality, documentation, and overall user experience! I have addressed most of your comments in different commits that listed above. Here is a summary of the major updates:

1. Enhanced the overall user guide and clarified the purpose of each example to help users determine which example to follow/adapt to best suits their research needs.
2. Added details on the example climate data, including explanations of raw CMIP6 versus ISIMIP-processed CMIP6 data and the required climate data format for `gaia`.
3. Provided additional explanations on both intermediate and final model outputs.
4. Updated the `crop_calendars` function to support more crop selections and included example code to guide users in modifying crop calendar data.
5. Improved the documentation of input data sources in the code.
6. Reduced the size of the example climate dataset to minimize download errors and improve testing efficiency.
7. Removed unnecessary files from the repository.

---

The followings are more detailed response to specific comments.

> Why do you recommend the use of WATCH climate observations? This is a very old dataset, based on an outdated ERA model, and you only need temperature and precipitation anyway.

We appreciate the reviewer’s feedback regarding the WATCH dataset. We used the WATCH dataset because this is the same climate data used in Waldhoff et al., 2020 study, which developed the core approach embedded in `gaia`. However, we agreed with the reviewer that the WATCH dataset may be considered outdated. We have revised the tutorial language to remind users the options to choose alternative historical climate datasets. Please see commit 8876557.

> It is odd that Example 1 under Run gaia! uses both Example 1 and Example 2 from the climate data section. If both are required, then these are not separate examples.

The historical cropland-weighted climate data used in both Example 1 and Example 2 is sourced from example-data-1. The key difference between Examples 1 and 2 lies in the projected climate data: Example 1 utilizes a small-sized, cropland-weighted future climate dataset if users want to do a quick test of gaia, whereas Example 2 relies on larger climate NetCDF files and would take longer time to run and it is for those who would like to reproduce similar experiment but with their own NetCDF data. To avoid redundant downloads of the historical data, we put all cropland-weighted climate data within example-data-1.
We have revised the Example Data section to make this clearer. Please see commit 8876557.

> Most research that might want to use gaia will not be at the country level. Is this possible?

Thanks for your question. Currently, gaia supports outputs at both the country level and the region-basin level (specifically for the Global Change Analysis Model GCAM). The methodology is limited to the country scale because the model relies on global crop yield data from FAOSTAT, which is typically available at the country scale. Additionally, gaia was designed primarily to support global multisector dynamics models and integrated assessment models, which generally operate at the country level or even coarser scales.

> `weighted_climate`: What happens if the weather data does not cover all countries?

Thank you for your question. gaia requires weather data with global coverage to get the best results. If the data does not have global coverage, the model will encounter missing weather information for certain crop types that may cause error and it will have less data for empirical fitting.

> Table 3: Why does the output produced by data_aggregation have a separate yield for each month? What is grow_season? I can guess the others, but they should be defined. Why does Table 3 have temp and precip if the weather data is intended for table 4?

Thanks for your question. The annual yield varies by country, crop, and year. The data is formatted to show the monthly information on growing season, precipitation and temperature, so the annual yield is repeatedly showing at the monthly time step. Table 3 is for historical information and Table 4 is for the future weather. We have added more explanations in the section for clarity. Please see commit 6e0153a.

> Table 4: What are the min and max calculated over? Is it over mean monthly temperature and total monthly precip?

Thanks for your question. The min, max, and mean temperature and precipitation are all calculated based on cropland-weighted monthly temperature and precipitation for each year. We added explanations for the outputs. Please see commit 6e0153a.

> Report the meaning of the output values for yield_shock_projection. I'm guessing it's the ratio of projected yields in each year to the projected yield in the base year.

Thank you. The yield shock is the fractional change in the yield of a given crop grown in a specific country in a future period compared to the baseline yield expected under stable climate conditions. We have added explanations of yield shock in this section. Please see commit b8b2ad5.

> Based on the User Guide, this strikes me as more as a repo repository than a tool. It is not clear that the provided functions can be used for a different project, and we are only given examples that are part of a single workstream.

Thank you for your comment. The primary goal of this model is to equip researchers with tools to estimate country-level yield shocks under different future climate forcings using a peer-reviewed methodology. Gaia is designed as a crop model with built-in capabilities to processes, clean, and format climate data, global crop data (e.g., yield, planting, and harvesting) and outputs.

A key advantage of gaia is that it simplifies complex processes for researchers who may not have extensive experience in data manipulation or model development. It provides seamless integration for visualizing results and processing outputs in a format compatible with different versions of GCAM, which has a broad user community.

In addition, the provided functions are modular and adaptable for different projects that require country level yield shocks. Users can customize the inputs and workflow to suit their specific research needs, such as configuring the model, fitting the model with different climate and socioeconomic scenarios, CO2 emission trajectories, adding more crops, select time series of interest, and specifying the output options.

>  I was trying to follow the vignette, but the second example download (to https://zenodo.org/records/13976521/files/gaia_example_climate.zip) has timed out on me 5 times. I suggest you offer an option for people to download the files outside of that function, or reduce the size.

Thank you for pointing that out. To make the example data more suitable for download, we have reduced the size of the climate NetCDF file by shortening the time coverage from 2015-2100 to 2015-2030. This adjustment reduces the ZIP file size from 1.4 GB to 261 MB. Users who would like to try the full 21st-century time series can use the alternative dataset, example-data-1. We have also offered two options to download the data in the user guide, including through download URL link and through gaia's `get_example_data` function.
