### 2025-02-20 5e2393c
Add more tests on yield_impact function
#13
