### 2025-02-20 JGCRI/gaia#13 @mengqi-z
Dear @jrising,

Thank you for your comments on the repository and the paper! We have addressed them in various commits listed above. Some comments are already addressed in other issues. The followings are our responses to some of the specific comments.

> The autogenerated files in man do not have useful descriptions. These should be understandable without users resorting to the vignette.

Thank you for your comments. The `man/*.Rd` files are for R function documentation for both internal and external functions or data, where the details, including usage, arguments, return values, for external ones are made available for users to check under the Reference tab https://jgcri.github.io/gaia/reference/index.html.

> The unit tests only capture the vignette process, rather than considering multiple possible uses for each function.

Thank you for your feedback. We have updated the tests to cover more potential other uses for each function. Now the test coverage is above 90% of the entire code base based on the Codecov.

> Also, the Zenodo archive mentions the use of STITCHES to prepare the climate data, but this is not discussed in the user guide.

STITCHES is a climate emulator designed to generate ESM-like climate datasets under new scenarios. The climate output from STITCHES maintains the same structure and format as CMIP6 data. To ensure consistency with ISIMIP’s data format, we then applied the BASD approach for bias adjustment and statistical downscaling to create the current example dataset. However, we realize that using a STITCHES-generated dataset in this example may cause unnecessary confusion. To improve clarity and align with ISIMIP standard datasets, we have replaced our example data with the ISIMIP’s CanESM5 climate data for the SSP245 scenario.

> I remain concerned that this is essentially still a reproduction repository, since it is not clear that it can be used in cases other than reproducing the results of the Waldhoff paper.

Thank you for your comment. While Gaia can certainly be used to reproduce the results of the Waldhoff paper, it is designed as a flexible model with many customizable features, including the core feature of estimating yield shocks using different climate data from various global climate models, scenarios, and realizations. Like many models, Gaia’s core methodology is peer-reviewed, and users have lots of flexibility in selecting inputs to project yield shocks under different conditions.
Beyond many strengths discussed in the paper, gaia offers many customization options for users to adapt it to their own research needs. Here are some options among others:
* Options to input different climate forcings.
* Select additional crops from the MIRCA2000 crop list.
* Options to use user-provided crop planting and harvesting calendar data.
* Options to use different CO2 trajectories.
* Incorporate different socioeconomic impacts (SSP1 – 5) in terms of calculating agricultural productivity change for GCAM.
* Options to choose different GCAM versions for creating GCAM required agricultural productivity change XML file.

> The repo includes several datasets, but I do not see documentation of those datasets.

Thank you for your comment. There are both internal or external datasets, where the main datasets from the package as well as the example dataset include climate data, FAOSTAT yield data, SAGE crop calendar data, and MIRCA2000 harvested area data. We have provided references, including relevant papers or website links, in the documentation to guide users to full dataset details. Additionally, we have added in-code documentation where these datasets are introduced to improve clarity.

> I look forward to trying to reproduce the vignette results when I can download the associated data.

We have updated to a new future climate data that only covers 2015-2030 to reduce the data size (the original data covers 2015-2100). If you have issues using gaia's function `get_example_data` for downloading, you can download the data through the URL link provided in the user guide.

> Why is RStudio considered a necessary installation step?

Thank you for your comment. We provide this step in case someone is new to R and doesn’t have R already installed.
