### 2025-02-20 0c979d2
Update vignette for correct display of equation
