### 2025-02-20 JGCRI/gaia#10 @jrising
This is just a quick comment to thank you for your responses. I can't go through them now, but I will respond in a couple weeks.
