### 2025-02-28 bc9a278
Update interpolation function to improve efficiency
- also brings XML generation first before diagnostics
