### 2025-02-28 30ad136
Ensure the APC data read in correctly
The row count column is removed
