### 2025-02-28 #7538 @mengqi-z
@cheginit - Thanks for the reminder!  I have made various updates to address the comments and have posted my detailed responses to the reviewer in the following issues:

- https://github.com/JGCRI/gaia/issues/10
- https://github.com/JGCRI/gaia/issues/11
- https://github.com/JGCRI/gaia/issues/13
