### 2025-03-07 JGCRI/gaia#4 @klau506
Hi! Some more comments below:

1. L20 (abstract). Detail MIRCA acronym, as done with CMIP6 or ISIMIP.
2. L53: GGCM acronym, detail using capital letters, as done in the other acronyms.
3. L72: do you refer to CMIP6? Specify
4. L72: to boost clarity, you might consider rephrasing point 1. Both sentences can be integrated into one.
5. L83: "by year": all the data you are mentioning is annual, so maybe remove, since it might generate confusion.
