### 2025-03-07 JGCRI/gaia#6 @klau506
Hi! Some more comments below!

1. In the [Index](https://jgcri.github.io/gaia/index.html)>Introduction, the first paragraph, provide a link or a reference for GCAM.
2. In the [Index](https://jgcri.github.io/gaia/index.html)>Contributing, the contributing link does not work.
3. In the [User guide](https://jgcri.github.io/gaia/articles/vignette.html)>Functions, same comments than in the paper issue
4. In the [User guide](https://jgcri.github.io/gaia/articles/vignette.html)>Example, add link or description for the ISIMIP protocols
5. In the [User guide](https://jgcri.github.io/gaia/articles/vignette.html)>Example, it is not clear what's "climate from CMIP6 model CanESM5 under SSP245 scenario". Is this part of the approach or the scenarios contained in the dataset? Please, clarify
6. In the [User guide](https://jgcri.github.io/gaia/articles/vignette.html)>Example>Example1, detail (by adding description and/or URL) the acronyms from this sentence for people not familiarized with these items: "the historical climate is from WATCH dataset and the projected climate is from CanESM5 global climate model under SSP245 scenario"
7. In the [User guide](https://jgcri.github.io/gaia/articles/vignette.html)>Example>Example1,> blue box, add a link to "Example 3", and it is not clear how to use the `coef_default` function. Provide a detailed step-by-step example
