### 2025-03-10 ac75a54
Update paper.md
For #4
