### 2025-03-10 JGCRI/gaia#10 @jrising
Thank you for your responses. Most of these are fine, but I still have the following concerns:

Re: "CMIP-ISIMIP format": The link directs one to the protocol for simulation formatting, not input climate formatting. Also, it is a 17 page document. If the goal to for users to be able to provide their own data, the specific requirements needed for gaia should be specified. Also, the new text "Processes CMIP climate NetCDF data in accordance with the ISIMIP simulation protocols" suggests that you are processing data in accordance with the protocols, not that the data is in accordance with the protocols.

Re: "Run gaia! Example 3": This is still confusing. You label it an example. Presumably, you intend for it to be an example of how to use gaia, so what is the goal of the example and how do I follow it?

Re: adding new crop calendars: Table 3 shows 0's for every row under the oil_palm column, because only the first 10 rows are included.

I don't see responses to the following questions:
 - Although the guide mentions in multiple places that inputs can be changed-- as long as they follow the same format-- the user guide does not particularly facilitate. The formats aren't provided, and it's often not clear how changes at one point will affect things later on. Someone just has to follow the single workstream and hope for the best.
 - The regression uses a log dependent variable, so a weighted average isn't theoretically justified. It's okay at higher resolutions, but at the country level, you end up with a log-sum-exp problem.
 - It should be clarified that projected climate data needs to be BCSD relative to the same observational dataset used in the regression.
 - Standard errors are reported in the regression, but I don't see how uncertainty is handled. Ignoring it wouldn't be consistent with the regression results.

Additional issues trying to run the vignette:
 - In the Run gaia! Example 1 yield_impact command,`base_year = 2015` does not end with a comma.
 - I get the following error when I run either Run gaia! Example 1 or Example 2:
```
...
column names: iso, wheat, maize, rice, sorghum, soybean, sunflower, root_tuber, cassava, sugarcane, sugarbeet, cotton, plant, harvest
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_wheat.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_maize.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_rice.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sorghum.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_soybean.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sunflower.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_root_tuber.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_cassava.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sugarcane.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sugarbeet.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_cotton.csv"
Error in data.table::fread(file, skip = 0, stringsAsFactors = FALSE, header = TRUE) :
  input= must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \n or \r
```
