### 2025-03-10 e8e6f61
Update user guide
For #6
