+++ data-raw/DATASET.R
@@ -126,8 +126,8 @@ usethis::use_data(agprodchange_ni_gcam7, overwrite = TRUE)
 # This serves as the default regression relationship if no historical climate data
 # is provided
 coef_default <- data.table::data.table()
-for(crop_name in mapping_mirca_sage$crop_mirca){
-  coef_crop <- gaia::input_data(folder_path = 'C:/WorkSpace/GCIMS/GCIMS_Yield/regression_analysis/output/20230706_0942/data_processed',
+for(crop_name in crop_mirca$crop_name[!is.na(crop_mirca$crop_sage)]){
+  coef_crop <- gaia::input_data(folder_path = 'C:/WorkSpace/github/test_scripts/gaia/output_test_climate/data_processed',
                                 input_file =  paste("reg_out_", crop_name, "_", fit_name, ".csv", sep = ""),
                                 skip_number = 0 )
   coef_default <- dplyr::bind_rows(
