### 2025-03-11 JGCRI/gaia#6 @mengqi-z
Dear @klau506,

Thanks for spotting those issues! I have update the user guides to address all the points above. Please see commit e8e6f61.
