### 2025-03-11 #7538 @klau506
Dear @cheginit,

The authors have thoroughly addressed all the comments, leading to significant improvements in the manuscript, package, and documentation. This is a valuable and versatile work with applications across various fields. Overall, I believe it is now ready for publication.

Let me know if there's anything else I can assist with!

Best,
Clàudia
