### 2025-03-11 f82ae3d
Add details of formatting climate NetCDF
For #10
