+++ vignettes/vignette.Rmd
@@ -126,7 +126,7 @@ brew install gdal
 The primary functionality of `gaia` is encapsulated in the `yield_impact` wrapper function, which executes the entire workflow from climate data processing to yield shock estimation. Users can also execute individual functions to work through the main steps of the process ([Figure 1](#figure1)).
 
 
-1. [`weighted_climate`](#weighted_climate): Processes CMIP6 daily or monthly climate NetCDF data in accordance with the [ISIMIP simulation protocols](https://www.isimip.org/protocol/preparing-simulation-files/ "Here we specifically require the bias-adjusted and statistically downscaled CMIP climate data following ISIMIP's protocols") and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type.
+1. [`weighted_climate`](#weighted_climate): Processes CMIP6 daily or monthly climate NetCDF data formatted in accordance with the [ISIMIP simulation protocols](https://www.isimip.org/protocol/preparing-simulation-files/ "Here we specifically require the bias-adjusted and statistically downscaled CMIP climate data following ISIMIP's protocols") (more details [here](#note3)) and calculates cropland-weighted precipitation and temperature at the country level, differentiated by crop type and irrigation type.
 
 2. [`crop_calendars`](#crop_calendars): Generates crop planting months for each country and crop based on crop calendar data [Sacks et al., (2010)](https://doi.org/10.1111/j.1466-8238.2010.00551.x).
 
@@ -172,7 +172,7 @@ The primary functionality of `gaia` is encapsulated in the `yield_impact` wrappe
 If you plan to use other climate observations or simulations that are not directly from ISIMIP, please ensure that you:
 
 * Bias-adjust and statistically downscale your data to 0.5-degree resolution. You can follow ISIMIP’s official ISIMIP3BASD v2.5.0 approach ([Lange, 2021](https://doi.org/10.5281/zenodo.4686991); [Lange, 2019](https://doi.org/10.5194/gmd-12-3055-2019)), or use the [basd software](https://jgcri.github.io/basd/) developed by the Pacific Northwest National Laboratory.
-* Format your data according to ISIMIP's guidelines for [preparing simulation files](https://www.isimip.org/protocol/preparing-simulation-files/) in terms of NetCDF headers, grid format, variables and dimensions, and the time axis.</p>
+* Format your data according to ISIMIP's guidelines for [preparing simulation files](https://www.isimip.org/protocol/preparing-simulation-files/) in terms of NetCDF headers, grid format, variables and dimensions, and the time axis. More tips can be found [here](#note3).</p>
 </div>
 
 <br />
@@ -271,6 +271,84 @@ tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_ssp245_2015
 
 ```
 
+If you have your own climate forcing data, please follow the guidance below to format your data to be compatible with `gaia`.
+
+<a name="note3"></a>
+<details>
+<summary>Tips for Formatting Climate Forcing NetCDF (Click to Expand the Content)</summary>
+
+<div class="alert alert-info">
+<h4 class="alert-heading">Formatting Your Climate Forcing Data for gaia</h4>
+  <p class="mb-0">
[…]
