### 2025-03-11 ffa688f
Rename Example 3 and revise the crop calendar modification codee
For #10

- Rename example 3 to be explore output to clarify the purpose of this section
- Update the example code provided for users to modify crop calendar, so that people can see the immediate changes decribed in Table 3
