### 2025-03-11 82dbc3d
Update yield_impact.R
