### 2025-03-12 cbb8962
Update output paths for example 1 and 2
For #10

Solve the error with writing processed example-data-2 into the example-data-1 folder.
