### 2025-03-12 JGCRI/gaia#10 @mengqi-z
Dear @jrising,

Thank you for your comments. We have made several updates, as demonstrated in the linked commits above. Below are our responses to your comments:

> Re: "CMIP-ISIMIP format": The link directs one to the protocol for simulation formatting, not input climate formatting. Also, it is a 17 page document. If the goal to for users to be able to provide their own data, the specific requirements needed for gaia should be specified. Also, the new text "Processes CMIP climate NetCDF data in accordance with the ISIMIP simulation protocols" suggests that you are processing data in accordance with the protocols, not that the data is in accordance with the protocols.

Thank you for your comment. The ISIMIP climate forcing data format shares similarities with the impact model simulation output format. Unfortunately, there is no dedicated ISIMIP page detailing the input climate forcing format, which is why we referenced the simulation protocols. We agree that this document is quite long. To address this, we have summarized the key formatting requirements compatible with `gaia` [here](https://jgcri.github.io/gaia/articles/vignette.html#note3). We've also updated the text to "Processes CMIP6 daily or monthly climate NetCDF data formatted in accordance with the ISIMIP simulation protocols."

> Re: "Run gaia! Example 3": This is still confusing. You label it an example. Presumably, you intend for it to be an example of how to use gaia, so what is the goal of the example and how do I follow it?

Thank you for pointing this out. This section is intended to explain the intermediate and final outputs from running `gaia`. The section is organized into subsections that align with each step of the workflow, corresponding to a function, for a clearer structure. Another purpose of this section is to provide guidance on advanced application of `gaia`, such as customizing the crop calendar to include new crops. To address your concern, we have renamed the section from "Example 3" to "Explore Outputs", and added a brief introductory description to clarify the purpose of this section.

> Re: adding new crop calendars: Table 3 shows 0's for every row under the oil_palm column, because only the first 10 rows are included.

Thank you for catching this! To keep the documentation relatively concise, we initially included only a limited number of rows for each table. We have now modified the example code to ensure that countries with newly added oil palm planting and harvesting months appear in the first 10 rows for demonstration purpose.

> Although the guide mentions in multiple places that inputs can be changed-- as long as they follow the same format-- the user guide does not particularly facilitate. The formats aren't provided, and it's often not clear how changes at one point will affect things later on. Someone just has to follow the single workstream and hope for the best.

Thank you. In our last revision, we made several updates to improve this, which are summarized in our update log. Specifically:

* We have provided detailed instructions on modifying inputs such as crop calendars.
* We have added more resources on the required format of climate forcing data.
* The documentation for each function includes detailed argument formats that users can follow, available [here](https://jgcri.github.io/gaia/reference/index.html).
* In the section Explore Outputs (previously Example 3), we have provided output tables so users can compare results and ensure they can reproduce the example.

We acknowledge that `gaia` has limitations regarding compatibility with different climate data sources, which is a common challenge as climate data can vary significantly across temporal and spatial scales. However, the User Guide provides full example data to demonstrate acceptable climate data formats and complete examples that allow users to verify input and output formats in case modifications are needed.

> The regression uses a log dependent variable, so a weighted average isn't theoretically justified. It's okay at higher resolutions, but at the country level, you end up with a log-sum-exp problem.

We appreciate your thoroughness. However, we are unclear about what the "log-sum-exponent" problem you are referencing. In our search, we only find the LSE "trick" referenced as one used to deal with very large or small numbers, which does not seem applicable in this case. Could you please provide a reference or more detailed description of the potential issue?

> It should be clarified that projected climate data needs to be BCSD relative to the same observational dataset used in the regression.

Thank you for your suggestion. We have added a note in the User Guide clarifying this requirement [here](https://jgcri.github.io/gaia/articles/vignette.html#note1).

> Standard errors are reported in the regression, but I don't see how uncertainty is handled. Ignoring it wouldn't be consistent with the regression results.

Thank you for your comment. Could you clarify which specific aspect of uncertainty you are referring to? The regression model follows standard practices by using the best estimates. We’d appreciate any further details on how uncertainty should be addressed in this context.

> In the Run gaia! Example 1 yield_impact command,base_year = 2015 does not end with a comma.

Thank you for catching this. We have updated the example to include the missing comma.

> I get the following error when I run either Run gaia! Example 1 or Example 2:

Thank you for catching this error. The issue is caused by saving the processed climate data from example-data-2 into the `weighted_climate/canesm5` folder of the downloaded example-data-1. This results in multiple weighted climate datasets with different time lengths being stored in the same folder, leading to issues when reading the climate data.

I have thoroughly updated the example codes to ensure separate output folders for Example 1 and Example 2. Please delete your current outputs and rerun the "Run gaia!" section from the beginning.
