### 2025-03-13 JGCRI/gaia#11 @jrising
Re: Dividing by 30 to get the month.
It's incorrect to divide by 30, and as a result you are misallocating days across months. Please revise it.

I don't see responses to gaia/R/data_aggregation.R line 30 and gaia/R/yield_impacts_functions.R line 439.

The other responses are fine. Thanks.
