### 2025-03-14 4ed2da8
Revise average days in a month value
For #11
