### 2025-03-14 fc71860
Add crop_calendar_file arg i ndata_aggregation
For #11

Add crop_calendar_file so if user use data_aggregation function by itself, they can still provide their own crop_calendar file
