+++ tests/testthat/helper.R
@@ -86,6 +86,7 @@ run_crop_calendars <- function(crop_calendar_file = NULL,
 run_data_aggregation <- function(data_dir = NULL,
                                  climate_model = climate_model_i,
                                  climate_scenario = climate_scenario_i,
+                                 crop_calendar_file = NULL,
                                  co2_hist = co2_hist_i,
                                  co2_proj = co2_proj_i,
                                  output_dir = output_dir_i){
@@ -97,6 +98,7 @@ run_data_aggregation <- function(data_dir = NULL,
                                    climate_impact_dir = climate_impact_dir_i,
                                    climate_model = climate_model,
                                    climate_scenario = climate_scenario,
+                                   crop_calendar_file = crop_calendar_file,
                                    co2_hist = co2_hist,
                                    co2_proj = co2_proj,
                                    output_dir = output_dir)
