+++ R/data_aggregation.R
@@ -7,6 +7,7 @@
 #' @param climate_model Default = NULL. String for climate model (e.g., 'CanESM5')
 #' @param climate_scenario Default = NULL. String for climate scenario (e.g., 'ssp245')
 #' @param historical_periods Default = NULL. Vector for years to subset from the historical climate data. If NULL, use the default climate data period
+#' @param crop_calendar_file Default = NULL. String for the path of the crop calendar file. If crop_calendar_file is provided, crop_select will be set to crops in crop calendar. User provided crop_calendar_file can include any crops MIRCA2000 crops: "wheat", "maize", "rice", "barley", "rye", "millet", "sorghum", "soybean", "sunflower", "root_tuber", "cassava", "sugarcane", "sugarbeet", "oil_palm", "rape_seed", "groundnuts", "pulses", "citrus", "date_palm", "grapes", "cotton", "cocoa", "coffee", "others_perennial", "fodder_grasses", "others_annual"
 #' @param co2_hist Default = NULL. Data table for historical CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
 #' @param co2_proj Default = NULL. Data table for projected CO2 concentration in columns [year, co2_conc]. If NULL, use built-in CO2 emission data
 #' @param output_dir Default = file.path(getwd(), 'output'). String for output directory
@@ -18,16 +19,33 @@ data_aggregation <- function(climate_hist_dir = NULL,
                              climate_model = "gcm",
                              climate_scenario = "rcp",
                              historical_periods = NULL,
+                             crop_calendar_file = NULL,
                              co2_hist = NULL,
                              co2_proj = NULL,
                              output_dir = file.path(getwd(), "output")) {
   message("Starting Step: data_aggregation")
 
   # read crop calendar output from crop_calendars function
-  crop_cal <- gaia::input_data(
-    folder_path = file.path(output_dir, "data_processed"),
-    input_file = "crop_calendar.csv"
-  )
+  if(is..null(crop_calendar_file)){
+
+    # Use default crop_calendar.csv created from the workflow
+    gaia::path_check(file.path(output_dir, "data_processed", "crop_calendar.csv"))
+    crop_cal <- gaia::input_data(
+      folder_path = file.path(output_dir, "data_processed"),
+      input_file = "crop_calendar.csv"
+    )
+
+  } else {
+
+    # Use user provided crop_calendar_file
+    gaia::path_check(crop_calendar_file)
+    crop_cal <- gaia::input_data(
+      folder_path = dirname(crop_calendar_file),
+      input_file = basename(crop_calendar_file)
+    )
+
+  }
+
 
   crop_select <- names(crop_cal)[!names(crop_cal) %in% c('iso', 'plant', 'harvest')]
 
