### 2025-03-14 c6f548d
Update data_aggregation.R
bug fix
