### 2025-03-17 JGCRI/gaia#13 @jrising
Thank you for the additional documentation and unit tests.

Re: autogenerated files in man. The following man entries do not have adequate information: agprodchange_ni_gcam6 , agprodchange_ni_gcam7 , clean_yield, climate_impact, co2_historical, co2_projection, coef_default, country_id, get_example_data, input_data, merge_data, output_data, pkg_example, yield_impact, and z_estimate. Some of these these have no description, and none of them have adequate text in the "details" section.

Re: documentation of datasets. Where is the documentation for [co2_historical.rda](https://github.com/JGCRI/gaia/blob/main/data/co2_historical.rda)? Also, what are the licenses for the dataset included in the repo?

The other responses are fine.
