+++ man/pkg_example.Rd
@@ -15,3 +15,4 @@ A string of path for files within the extras folder of gaia package
 \description{
 list example file paths
 }
+\keyword{internal}
