### 2025-03-19 JGCRI/gaia#13 @mengqi-z
Hi @jrising,

Thank you for looking into these details. Regarding your new comments:

> Re: autogenerated files in man. The following man entries do not have adequate information: agprodchange_ni_gcam6 , agprodchange_ni_gcam7 , clean_yield, climate_impact, co2_historical, co2_projection, coef_default, country_id, get_example_data, input_data, merge_data, output_data, pkg_example, yield_impact, and z_estimate. Some of these these have no description, and none of them have adequate text in the "details" section.

Thank you for your comments.  Base R provides a standard way of documenting a package where each function is documented in a topic, an `.Rd` file (“R documentation”) in the `man/` directory. These files are generated by the `roxygen2` package. If the function or data are defined as "external", then their documentation are accessible by users through [Package Index Page](https://jgcri.github.io/gaia/reference/index.html). However, not all the functions and data are exported externally for use by users.

- Some of those entries are designed as **internal functions**, including `clean_yield`, `climate_impact`, `input_data`, `merge_data`, `output_data`, and `z_estimate`. Those functions are internal helpers and are not supposed to be used directly by users. Their information can only be found within the package code.
- Some are **external functions** that are exported from the package to be used by users, including `get_example_data` and `yield_impact`. Those functions have detailed documentations which can be found in [Package Index Page](https://jgcri.github.io/gaia/reference/index.html).
- The rest are the package **external data** that are exported to be accessible by users, including `co2_historical`, `co2_projection`, `coef_default`, and `country_id`. You can find their details including source, format, and example usage in [Package Index Page](https://jgcri.github.io/gaia/reference/index.html). We have added more information as well to the external data.

> Re: documentation of datasets. Where is the documentation for [co2_historical.rda](https://github.com/JGCRI/gaia/blob/main/data/co2_historical.rda)? Also, what are the licenses for the dataset included in the repo?

Thank you. `data/*.rda` files are package data files, which are essentially generated from `usethis` package to store data in an R object. As mentioned above, if the package data is made external, then their documentation can be accessed in the [Package Index Page](https://jgcri.github.io/gaia/reference/index.html). In the case of `co2_historical.rda`, it is the R object for the external data `gaia::co2_historical`, which is documented [here](https://jgcri.github.io/gaia/reference/co2_historical.html). All the data included in the repo will follow the [BSD_3_clause](https://opensource.org/licenses/BSD-3-Clause).
