### 2025-03-20 JGCRI/gaia#11 @mengqi-z
Hi @jrising,

> Re: Dividing by 30 to get the month.
It's incorrect to divide by 30, and as a result you are misallocating days across months. Please revise it.

We appreciate your thoughtful suggestions. However, planting dates can vary widely and often extend across two months. For example, in Illinois, farmers may plant corn in April in some years and in May in others. There is no way to know which years they did what planting periods, and it is unavoidable to approximate the planting month. To address this, we take the mean of the planting start and planting end dates from the data and divide by 30 to estimate the planting month. In response to your comment, we have adjusted our calculation to use 30.4, which more accurately reflects the average number of days in a month over a year. Please see commit 4ed2da8.

> I don't see responses to gaia/R/data_aggregation.R line 30 and gaia/R/yield_impacts_functions.R line 439.

Thank you for catching this. We have `added crop_calendar_file` argument in `data_aggregation` function to enable the ability for users to provide their own crop calendar file path if they plan to use this function individually. Please see commit fc71860.

Regarding the `gaia/R/yield_impacts_functions.R` comment, we are still in the progress of testing the new approach. We will update you once it's ready.
