### 2025-03-21 eb8ff71
More improvements on the function/data documentation
