+++ man/co2_projection.Rd
@@ -8,13 +8,13 @@
 R data frame
 }
 \source{
-GCAM7 reference CO2 concentration; obtained from default GCAM7 Reference scenario CO2 concentration projection
+https://gmd.copernicus.org/articles/12/677/2019/gmd-12-677-2019.html
 }
 \usage{
 co2_projection
 }
 \description{
-co2_projection
+Projected global annual mean CO2 concentration based on GCAM7 Reference scenario, which is aligned with RCP7.0. Obtained from running default GCAM7 Reference scenario. This is the default CO2 projection used in gaia. User can update the CO2 concentration if preferred by using the argument co2_proj from yield_impact function.
 }
 \examples{
 \dontrun{
