+++ man/agprodchange_ni_gcam7.Rd
@@ -8,13 +8,13 @@
 R data frame
 }
 \source{
-GCAM7 ag productivity change reference data https://jgcri.github.io/gcam-doc/inputs_supply.html
+https://jgcri.github.io/gcam-doc/inputs_supply.html
 }
 \usage{
 agprodchange_ni_gcam7
 }
 \description{
-agprodchange_ni_gcam7
+Agricultural productivity change from GCAM7 Reference scenario
 }
 \examples{
 \dontrun{
