### 2025-03-21 JGCRI/gaia#11 @mengqi-z
Hi @jrising - Thank you for your comments on the standard error term in  `gaia/R/yield_impacts_functions.R`. We have addressed your comment by implementing robust standard errors clustered by country. Please see commit bfa9fc5.
