### 2025-03-25 JGCRI/gaia#10 @jrising
Re: log-sum-exp problem:

The problem comes from the fact that there is no pont-level data-generating process that is consistent with a log dependent variable regression. When you use different sized regions (countries), you are assuming that there is. If your regression is log(y) = f(T), but y = \sum_i w_i y_i, summing production over regions in a country, then you get log(y) = log(\sum_i w_i exp(f(T_i))). According to the log-sum-exp relationship (https://en.wikipedia.org/wiki/LogSumExp), that's approximately equal to f(T_j), where j is the largest yielding region, not f(\sum_i w_i T_i) like you use.

Re: Standard errors are reported in the regression

Standard practice when projecting econometric results under future climate is to report uncertainty around the final projection, as driven by uncertainty in the underlying coefficients. The normal way to do this is by taking Monte Carlo draws from the multivariate normal distribution of all parameters (using the variance-covariance matrix) and projecting each set of parameters, which then gives you an empirical distribution of final outcomes.

Further problems with the examples:

I was able to get the Run gaia! example 2 working with data from example 1. It may be confusing to users that the example 1 data doesn't map to the example 1 running.

I deleted all my data, updated the package, and ran the code from scratch, but when I try to do Run gaia! example 1 I get the error:
```
Error in data.table::fread(file, skip = 0, stringsAsFactors = FALSE, header = TRUE) :
  input= must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \n or \r
```
