### 2025-03-25 JGCRI/gaia#10 @mengqi-z
Hi @jrising,

Re: log-sum-exp problem:
Thank you for your clarification. We believe there may be a misunderstanding regarding our approach. The empirical model conducts regression at the country level using country-level yield observation data, along with cropland area-weighted country-level precipitation and temperature. Importantly, there is no aggregation from point-level yield or production data to country level.

Re: Standard errors are reported in the regression
Thank you for the explanation. The uncertainty analysis is beyond the purpose of the software. However, `gaia` provides all the information in the output files `reg_out_[crop]_xxx.csv` for users to do a Monte Carlo analysis if desired.

Re: Further problems with the examples
Thank you for testing. Could you please confirm that you have also deleted all example climate data? I suspect that your previous runs following Example 1 may have generated climate output files in the downloaded climate data folder (e.g., weighted_climate/canesm5), which could be causing the errors you encountered.
