+++ vignettes/vignette.Rmd
@@ -225,14 +225,14 @@ climate_impact_dir <- file.path(data_dir, 'canesm5')
 <a name="note2"></a>
 <div class="alert alert-info">
   <h4 class="alert-heading">NOTE!</h4>
-  <p class="mb-0"> The default example historical climate data provided in [example data 1](#example-data-1-quick-start-dataset) is based on WATCH climate observations ([Weedon et al., (2011)](https://doi.org/10.1175/2011JHM1369.1)), which were used in [Waldhoff et al., (2020)](https://doi.org/10.1088/1748-9326/abadcb) for empirical model fitting. If you do not intend to modify any assumptions or historical climate forcing for empirical model fitting, you can use the default regression model in `gaia`, which is pre-fitted with WATCH historical climate data, by setting the argument `use_default_coeff = TRUE`. After loading the `gaia` package, type `coef_default` to view the parameters of the default pre-fitted model. Alternatively, if you prefer to use different historical climate data other than the one from [example data 1](#example-data-1-quick-start-dataset), you can use `weighted_climate` function to generate cropland-weighted historical monthly precipitation and temperature following the [weighted_climate instruction](#weighted_climate) under [Example 3](#example-3).</p>
+  <p class="mb-0"> The default example historical climate data provided in [example data 1](#example-data-1-quick-start-dataset) is based on WATCH climate observations ([Weedon et al., (2011)](https://doi.org/10.1175/2011JHM1369.1)), which were used in [Waldhoff et al., (2020)](https://doi.org/10.1088/1748-9326/abadcb) for empirical model fitting. If you do not intend to modify any assumptions or historical climate forcing for empirical model fitting, you can use the default regression model in `gaia`, which is pre-fitted with WATCH historical climate data, by setting the argument `use_default_coeff = TRUE`. After loading the `gaia` package, type `coef_default` to view the parameters of the default pre-fitted model. Alternatively, if you prefer to use different historical climate data other than the one from [example data 1](#example-data-1-quick-start-dataset), you can use `weighted_climate` function to generate cropland-weighted historical monthly precipitation and temperature following the [weighted_climate instruction](#weighted_climate) under [Explore Outputs section](#example-3).</p>
 </div>
 
 <br />
 
 ## Example Data 2: Raw Climate Data
 
-Download the example data 2 using the instructions below. This dataset includes future global monthly precipitation and temperature from 2015 to 2030 at 0.5-degree resolution, projected by CanESM5 global climate model under SSP245 scenario. To use a different climate model or scenario, users can provide their own climate data in NetCDF format. Note that `gaia` adheres to the ISIMIP climate data format as its standard, so your NetCDF files should be formatted accordingly if the data is not directly downloaded from ISIMIP. More details on ISIMIP format can be found in [weighted_climate](#weighted_climate) section.
+Download the example data 2 using the instructions below. This dataset includes future global monthly precipitation and temperature from 2015 to 2030 at 0.5-degree resolution, projected by CanESM5 global climate model under SSP245 scenario. To use a different climate model or scenario, users can provide their own climate data in NetCDF format. Note that `gaia` adheres to the ISIMIP climate data format as its standard, so your NetCDF files should be formatted accordingly if the data is not directly downloaded from ISIMIP. More details on ISIMIP format can be found in [Tips for Formatting Climate Forcing NetCDF](#note3) section.
 
 There are two options to download the data:
 
@@ -275,7 +275,7 @@ If you have your own climate forcing data, please follow the guidance below to f
 
 <a name="note3"></a>
 <details>
-<summary>Tips for Formatting Climate Forcing NetCDF (Click to Expand the Content)</summary>
[…]
