### 2025-03-27 a71f409
Improve speed in APC calculation and plotting
