### 2025-03-31 JGCRI/gaia#11 @jrising
Thank you for the updates. I usually use (365.25 / 12) rather than 30.4, but I think this is fine.
