### 2025-03-31 JGCRI/gaia#13 @jrising
The entries in `man/` for your internal functions are visible to users, even though they aren't on the index page. If I search for `merge`, I see from `gaia` both a `merge_data` and a `data_merge` function. It is not unreasonable to expect that users will find the `merge_data` function, and find it confusing. Presumably, you would want to mark functions like `merge_data` with `@ noRd` and not (as currently done) with `@export`.

For external functions like `get_example_data`, how should I know when and how to use this function based on the help page? The only information is the description "Download and unzip example data with download URL". Grammar aside, the only knowledge I have of this function is from the vignette, and I would be very interested in knowing what kinds of example data it can download. If it is a function only designed to download the two datasets mentioned in the user guide, that's fine, but it should be documented as such.

For `co2_historical` and `co2_projection`, units would be a minimum bar. For `coef_default`, the regression is needed. For `country_id`, I don't know where these IDs are used. I can see in the User Guide that the file produced by `weighted_climate` uses these, but that's not documented in either the `country_id` or `weighted_climate` help pages.
