### 2025-04-01 a48b189
Improve data and function documentation
#13

- Remove .Rd files for most internal functions
- Improve documentation for external functions
