+++ R/data.R
@@ -54,7 +54,7 @@
 
 #' coef_default
 #'
-#' Regression coefficients generated by [`broom::tidy()`](https://broom.tidymodels.org/) for the regression model using gaia's example historical climate data, including estimate, standard error, statistic, p-value, for the 17 default crops supported by gaia.
+#' Regression coefficients generated by \href{https://broom.tidymodels.org/}{`broom::tidy()`} for the regression model using gaia's example historical climate data, including estimate, standard error, statistic, p-value, for the 17 default crops supported by gaia.
 #'
 #' @source Regression coefficients derived by using the default historical WATCH climate data
 #' @format R data frame
@@ -67,7 +67,7 @@
 
 #' country_id
 #'
-#' Mapping of country ID, country name, and country ISO codes. This table can be used to identify countries based on their IDs in weighted climate files generated by the `gaia::weighted_climate` function. For mode details, see the [User Guide](https://jgcri.github.io/gaia/articles/vignette.html#weighted_climate).
+#' Mapping of country ID, country name, and country ISO codes. This table can be used to identify countries based on their IDs in weighted climate files generated by the `gaia::weighted_climate` function. For mode details, see the \href{https://jgcri.github.io/gaia/articles/vignette.html#weighted_climate}{User Guide}.
 #'
 #' @source country ID and name mapping
 #' @format R data frame
