+++ R/yield_regression.R
@@ -1,6 +1,6 @@
 #' yield_regression
 #'
-#' Performs regression analysis fitted with historical annual crop yields, growing season monthly temperature and precipitation, CO2 concentrations, and GDP per capita. The default econometric model applied in gaia is from Waldhoff et al., (2020). User can specify alternative formulas that are consistent with the data processed in data_aggregation.
+#' Performs regression analysis fitted with historical annual crop yields, growing season monthly temperature and precipitation, CO2 concentrations, and GDP per capita. The default econometric model applied in gaia is from \href{https://www.doi.org/10.1088/1748-9326/abadcb}{Waldhoff et al., (2020)}. User can specify alternative formulas that are consistent with the data processed in `data_aggregation`.
 #'
 #' @param formula Default = NULL. String for regression formula
 #' @param diagnostics Default = TRUE. Logical for performing diagnostic plot
