### 2025-04-02 JGCRI/gaia#10 @jrising
Re: log-sum-exp problem:
That is what I understood. To be clear, though, crop yields are generated by a point-level (field-level) data-generating process. Anyway, the question then is, why did you do cropland area-weighted weather variables? The theoretical justification for that, when using a linear dependent variable, is because you have a point-level data generating process. With a log dependent variable, you are (I believe) using an approach that is both empirical biased and theoretically incorrect. In any case, you are saying you have no theoretical justification, so I see two options. You could run a hind-casting empirical test (probably outside the scope of your work here publishing a software package) or include a "limitations" section of the user guide where you note that only crop-weighted weather is allowed.

Re: Standard errors are reported in the regression:
Okay, this can also be included in a "limitations" section.

Re: Further problems with the examples:
This time I deleted all folders in the `gaia_output` directory before running the code. Here is output from the whole process:
```
> library(gaia)
>
> output_dir <- '~/research/reviews/gaia/gaia_output'
> data_dir <- gaia::get_example_data(
+   download_url = 'https://zenodo.org/records/14888816/files/weighted_climate.zip?download=1',
+   data_dir = output_dir
+ )
Starting download.
trying URL 'https://zenodo.org/records/14888816/files/weighted_climate.zip?download=1'
Content type 'application/octet-stream' length 47798548 bytes (45.6 MB)
==================================================
downloaded 45.6 MB

Download complete.
Starting unzip.
Data unzipped to: ~/research/reviews/gaia/gaia_output/weighted_climate
>
> climate_hist_dir <- file.path(data_dir, 'climate_hist')
> climate_impact_dir <- file.path(data_dir, 'canesm5')
>
> data_dir <- gaia::get_example_data(
+   download_url = 'https://zenodo.org/records/14888816/files/gaia_example_climate.zip?download=1',
+   data_dir = output_dir
+   )
Starting download.
trying URL 'https://zenodo.org/records/14888816/files/gaia_example_climate.zip?download=1'
Content type 'application/octet-stream' length 261537964 bytes (249.4 MB)
==================================================
downloaded 249.4 MB

Download complete.
Starting unzip.
Data unzipped to: ~/research/reviews/gaia/gaia_output/gaia_example_climate
>
> pr_projection_file <- file.path(data_dir, 'pr_monthly_canesm5_w5e5_ssp245_2015_2030.nc')
> tas_projection_file <- file.path(data_dir, 'tas_monthly_canesm5_w5e5_ssp245_2015_2030.nc')
>
> gaia::yield_impact(
+   pr_hist_ncdf = NULL,                    # path to historical precipitation NetCDF file (must follow ISIMIP format); only if you wish to use your own historical precipitation observation
+   tas_hist_ncdf = NULL,                   # path to historical temperature NetCDF file (must follow ISIMIP format); only if you wish to use your own historical temperature observation
+   pr_proj_ncdf = pr_projection_file,      # path to future projected precipitation NetCDF file (must follow ISIMIP format)
+   tas_proj_ncdf = tas_projection_file,    # path to future projected temperature NetCDF file (must follow ISIMIP format)
+   timestep = 'monthly',                   # specify the time step of the NetCDF data (monthly or daily)
+   climate_hist_dir = climate_hist_dir,    # path to the folder that holds cropland weighted historical climate observations
+   historical_periods = c(1960:2001),      # vector of historical years selected for fitting
+   climate_model = 'canesm5',              # label of climate model name
+   climate_scenario = 'ssp245',            # label of climate scenario name
+   member = 'r1i1p1f1',                    # label of ensemble member name
+   bias_adj = 'w5e5',                      # label of climate data for bias adjustment for the global climate model (GCM)
+   cfe = 'no-cfe',                         # label of CO2 fertilization effect in the formula (default is no CFE)
+   gcam_version = 'gcam7',                 # output is different depending on the GCAM version (gcam6 or gcam7)
+   use_default_coeff = FALSE,              # set to TRUE when there is no historical climate data available
+   base_year = 2015,                       # GCAM base year
+   start_year = 2015,                      # start year of the projected climate data
+   end_year = 2030,                        # end year of the projected climate data
+   smooth_window = 20,                     # number of years as smoothing window
+   co2_hist = NULL,                        # historical annual CO2 concentration. If NULL, will use default value
+   co2_proj = NULL,                        # projected annual CO2 concentration. If NULL, will use default value
+   crop_select = NULL,                     # set to NULL for the default crops
+   diagnostics = TRUE,                     # set to TRUE to output diagnostic plots
+   output_dir = output_dir                 # path to the output folder
+ )
No precipitation files provided. Skipping.
No temperature files provided. Skipping.
Processing irc_crop01
Processing irc_crop02
Processing irc_crop03
...
Processing rfc_crop26
The function weighted_climate is complete.
Processing: ~/research/reviews/gaia/gaia_output/gaia_example_climate/pr_monthly_canesm5_w5e5_ssp245_2015_2030.nc
Processing: ~/research/reviews/gaia/gaia_output/gaia_example_climate/tas_monthly_canesm5_w5e5_ssp245_2015_2030.nc
Processing irc_crop01
Processing irc_crop02
Processing irc_crop03
...
Processing rfc_crop26
The function weighted_climate is complete.
Starting Step: crop_calendars
[1] "Final selected crops are: wheat, maize, rice, sorghum, soybean, sunflower, root_tuber, cassava, sugarcane, sugarbeet, cotton"
Starting Step: data_aggregation
Reading data... crop_calendar.csv
789 rows
14 columns
column names: iso, wheat, maize, rice, sorghum, soybean, sunflower, root_tuber, cassava, sugarcane, sugarbeet, cotton, plant, harvest
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_wheat.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_maize.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_rice.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sorghum.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_soybean.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sunflower.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_root_tuber.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_cassava.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sugarcane.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_sugarbeet.csv"
[1] "Merged historical yield data saved to: ~/research/reviews/gaia/gaia_output/data_processed/historic_vars_cotton.csv"
Error in data.table::fread(file, skip = 0, stringsAsFactors = FALSE, header = TRUE) :
  input= must be a single character string containing a file name, a system command containing at least one space, a URL starting 'http[s]://', 'ftp[s]://' or 'file://', or, the input data itself containing at least one \n or \r
In addition: There were 50 or more warnings (use warnings() to see the first 50)
```
