### 2025-05-05 JGCRI/gaia#10 @mengqi-z
Hi @jrising - Just wanted to let you know that I am back to work and we are working on addressing your last few comments. For the error messages you ran into, I believe I have addressed those issues in the latest updates. These updates will read the correct input files based on the model, scenario, years you have selected. Could you please make sure you have pulled all the latest changes to your cloned repository, start with a new empty folder and rerun the examples again? Thanks!
