### 2025-05-08 JGCRI/gaia#10 @jrising
Great, I have now been able to run Example 1! Just a couple of things I noticed.

1. I get the following warning messages at the end of the process:
```
Warning messages:
1: In system2("java", args, stdout = TRUE, stderr = TRUE) :
  running command ''java' -cp '/Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/gcamdata/extdata/ModelInterface/CSVToXML.jar' -Xmx2g ModelInterface.ModelGUI2.csvconv.CSVToXMLMain /var/folders/rl/p7xkj5n56gb2hxfnwbc696x80000gn/T//RtmpdMK4I0/file17afa6fa64fbe '/Library/Frameworks/R.framework/Versions/4.4-arm64/Resources/library/gcamdata/extdata/mi_headers/ModelInterface_headers.txt' '~/research/reviews/gaia/gaia_output/gcam7_agprodchange_no-cfe/agyield_impact_canesm5_r1i1p1f1_w5e5_ssp245.xml' 2>&1' had status 1
2: In gcamdata::run_xml_conversion(.) :
  The operation couldn’t be completed. Unable to locate a Java Runtime.Please visit http://www.java.com for information on installing Java.
```

2. The data in the `figures` folder all looks plausible, except for `annual_projected_climate_impacts_canesm5_ssp245_cassava_fit_lnyield_mmm_quad_noco2_nogdp.pdf` and `smooth_projected_climate_impacts_canesm5_ssp245_cassava_fit_lnyield_mmm_quad_noco2_nogdp.pdf`. These both flat lines at 1:
