### 2025-06-10 18c382c
Fix the cassava crop calendar and add limitation section
