+++ vignettes/vignette.Rmd
@@ -551,7 +551,7 @@ knitr::kable(input_climate[1:12],
 
 ---
 
-The `crop_calendars` function calculates crop calendars based on the SAGE global crop planting and harvesting dates data ([Sacks et al., 2010](https://doi.org/10.1111/j.1466-8238.2010.00551.x)).  The crop calendar outlines the planting and harvesting months for each predefined crop across all global countries.`gaia` uses MIRCA2000 crop category ([Portmann et al., 2010](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2008gb003435)) as standard crop types. The code chunk below runs `crop_calendars` for the default crops in `gaia`: `cassava`, `cotton`, `maize`, `rice`, `root_tuber`, `sorghum`, `soybean`, `sugarbeet`, `sugarcane`, `sunflower`, `wheat.`
+The `crop_calendars` function calculates crop calendars based on the SAGE global crop planting and harvesting dates data ([Sacks et al., 2010](https://doi.org/10.1111/j.1466-8238.2010.00551.x)). The crop calendar outlines the planting and harvesting months for each predefined crop across all global countries.`gaia` uses MIRCA2000 crop category ([Portmann et al., 2010](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2008gb003435)) as standard crop types. The code chunk below runs `crop_calendars` for the default crops in `gaia`: `cassava`, `cotton`, `maize`, `rice`, `root_tuber`, `sorghum`, `soybean`, `sugarbeet`, `sugarcane`, `sunflower`, `wheat.`
 
 
 ```{r eval=F, echo=T}
@@ -566,6 +566,8 @@ crop_cal <- crop_calendars(output_dir = output_dir)
 crop_cal
 
 ```
+
+Please note that we have excluded countries with `cassava` planting seasons longer than two months, based on SAGE data. This is because extended planting periods from SAGE data typically indicate multiple planting and harvesting cycles, which makes it difficult to identify yield for each individual harvest.
 
 **Outputs of the function:** The function returns a data frame of crop calendars. It also writes the output in `output_dir/data_processed/crop_calendar.csv`  
 
@@ -938,5 +940,26 @@ gcam_apg <- gcam_agprodchange(data = out_yield_shock,
 <p align="left" style="font-size:14px;"> **Figure 4:** Agricultural productivity growth rate for soybean. Each line under a region indicates the APG for the region-basin intersection. </p>
 
 <br />
+
+# Limitations
+
+---
+
+The current version of `gaia` has several limitations that users should be aware of.
+
+## Weather Data Aggregation and Theoretical Consistency
+
+In the current implementation, `gaia` uses cropland area-weighted averages of weather variables (e.g., temperature and precipitation) to construct country-level climate inputs. While this aggregation method is commonly used in empirical crop yield modeling, its theoretical justification is more robust when the dependent variable (e.g., yield) is in linear form, as it can be interpreted as a consistent approximation of a field-level data-generating process.
+
+However, in `gaia`, the dependent variable is specified as the natural logarithm of yield, which allows us to project the future yield shocks based on weather only, allowing modelers to use their own assumptions about technological improvements driving yield improvements. In this context, the use of cropland-weighted average climate inputs might lead to theoretical inconsistency. Specifically, if yield is assumed to be generated by an underlying nonlinear function of weather at the field level (e.g., $$log(y)=f(T,P)$$), then aggregating climate variables before applying the function is not equivalent to aggregating yield outcomes after applying the function at the field level. This is a known issue related to the log-sum-exp transformation.
+
[…]
