+++ pkgdown.yml
@@ -1,6 +1,6 @@
 pandoc: 3.1.11
-pkgdown: 2.1.1
+pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   vignette: vignette.html
-last_built: 2025-04-01T20:57Z
+last_built: 2025-06-10T18:24Z
