### 2025-06-10 JGCRI/gaia#10 @mengqi-z
Hi @jrising,

Apologies for the delayed response. I've been tied up with other work commitments over the past few weeks.

RE:
> I get the following warning messages at the end of the process

Thank you for pointing this out. The warning messages you encountered originate from one of `gaia`'s dependencies, `gcamdata`, which is used within `gaia` to generate standard XML output files for the Global Change Analysis Model (GCAM), if requested.

Based on similar cases, I believe these warnings are typically caused by running a 32-bit version of Java. Could you confirm which version of Java you have installed, and whether you might have multiple versions on your system? Ensuring that you're using a 64-bit version of Java should resolve the issue.

You can check your Java version with a command in R console:
```R
> system2("java", "-version")
openjdk version "14.0.1" 2020-04-14
OpenJDK Runtime Environment (build 14.0.1+7)
OpenJDK 64-Bit Server VM (build 14.0.1+7, mixed mode, sharing)
```
If the version is 32-bit, then I would recommend updating your JAVA according to https://openjdk.org.

RE:
> The data in the figures folder all looks plausible, except for:
annual_projected_climate_impacts_canesm5_ssp245_cassava_fit_lnyield_mmm_quad_noco2_nogdp.pdf and smooth_projected_climate_impacts_canesm5_ssp245_cassava_fit_lnyield_mmm_quad_noco2_nogdp.pdf. These both flat lines at 1

Thank you for catching this issue. After some tests, we decided to omit the countries with planting seasons longer than two months for cassava. This is because some countries have long planting season for cassava based on the SAGE data, which indicates that there are multiple harvests. This makes it difficult to identify yield for each individual harvest and could lead to parameters being insignificant in the regression model, so that they are not used for yield shock projections. We’ve added a note in the [documentation](https://jgcri.github.io/gaia/articles/vignette.html#crop_calendars) to explain this cassava-specific adjustment.

RE:
> log-sum-exp problem and  Standard errors are reported in the regression

Thank you for the previous discussions on this. We have included a ["Limitations" section](https://jgcri.github.io/gaia/articles/vignette.html#limitations) in the documentation.
