+++ man/coef_default.Rd
@@ -14,7 +14,7 @@ Regression coefficients derived by using the default historical WATCH climate da
 coef_default
 }
 \description{
-Regression coefficients generated by [`broom::tidy()`](https://broom.tidymodels.org/) for the regression model using gaia's example historical climate data, including estimate, standard error, statistic, p-value, for the 17 default crops supported by gaia.
+Regression coefficients generated by \href{https://broom.tidymodels.org/}{`broom::tidy()`} for the regression model using gaia's example historical climate data, including estimate, standard error, statistic, p-value, for the 17 default crops supported by gaia.
 }
 \examples{
 \dontrun{
