+++ man/country_id.Rd
@@ -14,7 +14,7 @@ country ID and name mapping
 country_id
 }
 \description{
-Mapping of country ID, country name, and country ISO codes. This table can be used to identify countries based on their IDs in weighted climate files generated by the `gaia::weighted_climate` function. For mode details, see the [User Guide](https://jgcri.github.io/gaia/articles/vignette.html#weighted_climate).
+Mapping of country ID, country name, and country ISO codes. This table can be used to identify countries based on their IDs in weighted climate files generated by the `gaia::weighted_climate` function. For mode details, see the \href{https://jgcri.github.io/gaia/articles/vignette.html#weighted_climate}{User Guide}.
 }
 \examples{
 \dontrun{
