+++ .github/workflows/test-coverage.yaml
@@ -12,7 +12,7 @@ jobs:
   test-coverage:
     runs-on: ubuntu-latest
     env:
-      GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
+      GITHUB_PAT: ${{ secrets.GAEA_PAT }}
 
     steps:
       - uses: actions/checkout@v4
