+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   vignette: vignette.html
-last_built: 2025-06-11T03:13Z
+last_built: 2025-06-11T16:29Z
