### 2025-06-27 JGCRI/gaia#10 @jrising
Okay, this is all resolved to my satisfaction.
