### 2025-07-10 2b00eca
Update author and version information
