+++ CITATION.cff
@@ -4,8 +4,8 @@ authors:
 - family-names: "Zhao"
   given-names: "Mengqi"
   orcid: "https://orcid.org/0000-0001-5385-2758"
-- family-names: "Waldhoff"
-  given-names: "Stephanie"
+- family-names: "Morris"
+  given-names: "Stephanie T."
   orcid: "https://orcid.org/0000-0002-8073-0868"
 - family-names: "Claudia"
   given-names: "Tebaldi"
@@ -22,18 +22,18 @@ preferred-citation:
     - family-names: "Zhao"
       given-names: "Mengqi"
       orcid: "https://orcid.org/0000-0001-5385-2758"
-    - family-names: "Waldhoff"
-      given-names: "Stephanie"
+    - family-names: "Morris"
+      given-names: "Stephanie T."
       orcid: "https://orcid.org/0000-0002-8073-0868"
-    - family-names: "Claudia"
-      given-names: "Tebaldi"
+    - family-names: "Tebaldi"
+      given-names: "Claudia"
       orcid: "https://orcid.org/0000-0001-9233-8903"
-    - family-names: "Abigail"
-      given-names: "Snyder"
+    - family-names: "Snyder"
+      given-names: "Abigail"
       orcid: "https://orcid.org/0000-0002-9034-9948"
     doi: "10.xxxxx/joss.xxxxx"
     journal: "Journal of Open Source Software"
     title: "gaia: An R package to estimate crop yield responses to temperature and precipitation"
     issue: x
     volume: y
-    year: 2024
+    year: 2025
