### 2025-07-14 9ee136d
Fix the average impact calculation to remove NA
