+++ pkgdown.yml
@@ -3,4 +3,4 @@ pkgdown: 2.1.3
 pkgdown_sha: ~
 articles:
   vignette: vignette.html
-last_built: 2025-07-10T16:34Z
+last_built: 2025-07-14T23:49Z
