### 2025-07-15 abec478
Update yield_impacts_functions.R
Update the rowMeans with na.rm
