### 2025-07-15 #7538 @cheginit
@mengqi-z Please address these issues in the paper, then release a new version with all fixes, if applicable, and provide me with the version number and its DOI. Thanks for your patience!

- [ ] Line 11-12: Consider: "using the response surfaces from an empirical econometric model developed and documented in Waldhoff et al. (2020)"
- [ ] Line 16-17: Should be: "bias-adjusted and statistically downscaled by **the** ISIMIP3BASD approach"
- [ ] Line 25-26: Should be: "that can aid exploration of crop **yield** responses" (remove 's' from yields)
- [ ] Line 40-41: Consider: "There is a need to include projections of inter-annual weather impacts on crop yields for multiple crops under a wide range of climate projections"
- [ ] Line 50: Should be: "The exploration of **the** effects of future climate change"
- [ ] Line 56-57: Should be: "making **extensions** to many crops and regions more challenging"
- [ ] Line 69-70: No hyphen needed. Should be: "decision **makers**"
- [ ] Line 72: Consider adding "an" before "example dataset": "with **an** example dataset"
- [ ] Line 77: Should be: `crop_calendars` (not "calenders")
- [ ] Line 85: Should be: "**Users** can specify alternative formulas" (not "User")
- [ ] Line 94: Consider: "**and** aggregates crops into GCAM commodities"
- [ ] Throughout paper: Inconsistent hyphenation of "country-level" vs "country level" - choose one style and maintain consistency
- [ ] Throughout paper: Inconsistent terminology between "crop yields" and "yield changes" - ensure consistent usage
