### 2025-07-15 32fa595
Update paper.md
