+++ paper/paper.md
@@ -22,39 +22,39 @@ affiliations:
    index: 1
  - name: Pacific Northwest National Laboratory, Joint Global Change Research Institute, College Park, MD, USA
    index: 2
-date: 2024
+date: 2025
 bibliography: paper.bib
 ---
 
 # Summary
 
-`gaia` is an open-source R package designed to estimate crop yield shocks in response to annual weather variations and CO~2~ concentrations at the country scale for 12 major crops. This innovative tool streamlines the workflow from raw climate data processing to projections of annual shocks to crop yields at the country level, using the response surfaces developed and documented in @Waldhoff_2020, an empirical econometric model that leverages historical weather, CO~2~, and crop yield data for robust empirical fitting for 12 crops. `gaia` uses these response surfaces with bias-corrected, gridded monthly temperature and precipitation projections (e.g., from the Coupled Model Intercomparison Project Phase 6 (CMIP6) [@Oneil_2016] climate data bias-adjusted and statistically downscaled by ISIMIP3BASD approach [@Lange_2019] in the Inter-Sectoral Impact Model Intercomparison Project (ISIMIP) [@Warszawski_2014]) to project shocks that can be applied to agricultural productivity changes at the country-level for use in multisectoral economic models. The historical and future projections use gridded, country-crop specific monthly growing season precipitation and temperature data, aggregated to the national level, weighted by cropland area derived from the global Monthly Irrigated and Rainfed Crop Areas around the year 2000 (MIRCA2000) dataset [@Portmann_2010]. These annual, country, crop-specific yield shocks can be aggregated to different definitions of regions, crop commodities, and time periods, as needed by specific multisectoral economic models. `gaia` serves as a lightweight, powerful tool that can aid exploration of crop yields responses under a broad range of future climate projections, enhancing human-Earth system analysis capabilities.
+`gaia` is an open-source R package designed to estimate crop yield shocks in response to annual weather variations and CO~2~ concentrations at the country scale for 12 major crops. This innovative tool streamlines the workflow from raw climate data processing to projections of annual shocks to crop yields at the country level, using the response surfaces from an empirical econometric model developed and documented in @Waldhoff_2020, which leverages historical weather, CO~2~, and crop yield data for robust empirical fitting for 12 crops. `gaia` uses these response surfaces with bias-corrected, gridded monthly temperature and precipitation projections (e.g., from the Coupled Model Intercomparison Project Phase 6 (CMIP6) [@Oneil_2016] climate data bias-adjusted and statistically downscaled by the ISIMIP3BASD approach [@Lange_2019] in the Inter-Sectoral Impact Model Intercomparison Project (ISIMIP) [@Warszawski_2014]) to project shocks that can be applied to agricultural productivity changes at the country level for use in multisectoral economic models. The historical and future projections use gridded, country-and-crop specific monthly growing season precipitation and temperature data, aggregated to the national level, and weighted by cropland area derived from the global Monthly Irrigated and Rainfed Crop Areas around the year 2000 (MIRCA2000) dataset [@Portmann_2010]. These annual, country, and crop-specific yield shocks can be aggregated to different definitions of regions, crop commodities, and time periods, as needed by specific multisectoral economic models. `gaia` serves as a lightweight, powerful tool that can aid exploration of crop yield responses under a broad range of future climate projections, enhancing human-Earth system analysis capabilities.
 
 
 # Statement of need
 
[…]
