+++ NEWS.md
@@ -1,4 +1,4 @@
-# gaia 1.0.0
+# gaia 2.0.1
 
 `gaia` is a powerful and user-friendly tool designed to estimate crop yield responses to climate impacts, enabling robust projections and streamlined workflows. Explore its full capabilities in the [User Guide](https://jgcri.github.io/gaia/articles/vignette.html).
 
@@ -8,8 +8,10 @@
 
 * **Empirical Econometric Modeling:** Estimate annual crop yield responses to climate impacts using an empirically grounded approach.
 * **Streamlined Workflow:** Simplify your analysis with a single wrapper function, `yield_impact`, for running models efficiently.
+* **Integrated Climate Data Processing:** Process widely-used ISIMIP-style climate NetCDF files without the need for external tools.
 * **Data-Driven Insights:** Leverage historical weather, CO2, and crop yield data for robust model fitting.
-* **Future Projections:** Project annual yield shocks for 12 crops using CMIP/ISIMIP bias-adjusted precipitation and temperature datasets. `gaia` supports both daily and monthly climate data formats.
+* **Future Projections:** Project annual yield shocks for 17 crops using CMIP/ISIMIP bias-adjusted precipitation and temperature datasets. `gaia` supports both daily and monthly climate data formats.
+* **Flexible Crop Coverage:** Estimate yield shocks for 17 major crops, with the option to add more crops by providing appropriate input data.
 * **Model Supports:** Designed to support [Global Change Analysis Model (GCAM)](https://github.com/JGCRI/gcam-core) as well as other non-GCAM modeling frameworks.
 
 
