+++ DESCRIPTION
@@ -1,7 +1,7 @@
 Package: gaia
 Title: gaia: An R package to estimate crop yield responses to temperature
     and precipitation
-Version: 2.0.0
+Version: 2.0.1
 Authors@R: c(
     person("Mengqi", "Zhao", , "mengqi.zhao@pnnl.gov", role = c("aut", "cre"),
            comment = c(ORCID = "0000-0001-5385-2758")),
