### 2025-07-15 403f036
Update paper.md
